### 2023-09-20 #5860 @editorialbot
[PRE REVIEW]: NEoST: A Python package for nested sampling of the neutron star equation of state
**Submitting author:** @drannawatts (<a href="http://orcid.org/0000-0002-1009-2354">Anna Watts</a>)
**Repository:** https://github.com/xpsi-group/neost
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.1
