### 2023-09-21 #5860 @drannawatts
> @editorialbot query scope
>
> @drannawatts — Thanks for your submission! Given the short development history of the code, I have submitted this pre-review for editorial review. This usually takes a week or two and I'll report back here as soon as I hear back from the rest of the editorial board.

The code has actually existed in private repositories for much longer (we've been using it in analysis and publications since 2019).  We moved the whole thing over to a new public repository with a clean history specifically because we wanted to finally make it public.
