### 2023-11-01 #6003 @editorialbot
[REVIEW]: NEoST: A Python package for nested sampling of the neutron star equation of state
**Submitting author:** @drannawatts (<a href="http://orcid.org/0000-0002-1009-2354">Anna Watts</a>)
**Repository:** https://github.com/xpsi-group/neost
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.0
