### 2023-11-02 #6003 @adonath
@drannawatts @ziatdinovmax @chaitanyaafl this is the review thread for the paper. All of our communications will happen here from now on. Thanks again for the submission and for agreeing to review!

The instructions for the review are given at the beginning of the page. I would like to emphasize again that reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention openjournals/joss-reviews#REVIEW_NUMBER so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
