### 2024-03-29 #6003 @drannawatts
Just checking in on this (since we're about to submit a new NICER results EoS paper using NEoST).  If there's anything you need from us please let us know!
