### 2024-04-16 #6003 @adonath
@tddesjardins You can find some more instructions for the review process here: https://github.com/openjournals/joss-reviews/issues/6003#issuecomment-1790661927 and here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
