### 2024-07-02 aaf26c2
Numpy 1.26 compatibility: use "density" instead of "normed" in numpy.histogramdd.
According to numpy documentation, these are identical but "normed" is deprecated".
	modified:   neost/PosteriorAnalysis.py
