+++ docs/source/conf.py
@@ -49,7 +49,6 @@ intersphinx_mapping = {'sphinx': ('http://www.sphinx-doc.org/en/master', None),
            'numpy': ('https://docs.scipy.org/doc/numpy', None),
            'cython': ('https://cython.readthedocs.io/en/latest', None),
            'emcee': ('https://emcee.readthedocs.io/en/latest', None),
-           'getdist': ('https://getdist.readthedocs.io/en/latest', None),
            'schwimmbad': ('https://schwimmbad.readthedocs.io/en/latest', None),
            'mpi4py': ('https://mpi4py.readthedocs.io/en/stable', None),
            'h5py': ('http://docs.h5py.org/en/latest', None),
