### 2024-07-02 a0ba578
scipy: "scipy.integrate.cumtrapz" was renamed to "scipy.integrate.cumulative_trapezoid" in scipy 1.6.0, and the old name was deprecated in scipy 1.14.0. Updated neost/eos/speedofsound.py to reflect this change. This breaks compatibility with old scipy < 1.6 (which may be broken already, not sure) 	modified:   neost/eos/speedofsound.py
