### 2024-07-02 6dcc582
Updated environment.yml to reflect the new dependency situation. 	modified:   environment.yml
