### 2024-07-02 84aa0fb
TOV solvers: improved the printouts for which version is used. Nothing is printed if the C solvers are used. A user warning is issued if the slower python solvers are used. 	modified:   neost/Star.py
