### 2024-07-02 c38c226
Store the radius corresponding to the maximum mass in a variable. 	modified:   neost/eos/base.py
