### 2024-07-02 33d5b8a
Added function to compute quantiles 	modified:   neost/PosteriorAnalysis.py
