### 2024-07-02 f880989
Added compute_table_data() function 	modified:   neost/PosteriorAnalysis.py
