### 2024-07-02 221672a
Added compute_prior_auxiliary_data() function 	modified:   neost/PosteriorAnalysis.py
