+++ neost/Likelihood.py
@@ -47,6 +47,7 @@ class Likelihood():
                                    'Lambda':star.tidal}
                 tmp = list(map(MassRadiusTidal.get, self.likelihood_params[i]))
                 like = self.likelihood_functions[i](tmp)
+                like = self.array_to_scalar(like)
                 likelihoods.append(like)
             else:
                 M2 = star.Mrot
@@ -58,11 +59,22 @@ class Likelihood():
                 point = numpy.array([self.chirp_masses[i], M2 / M1,
                                      MTspline(M1), star.tidal])
                 like = self.likelihood_functions[i](point)
+                like = self.array_to_scalar(like)
                 likelihoods.append(like)
 
         like_total = numpy.prod(numpy.array(likelihoods))
-        # print('lnlike is', like_total)
         if like_total == 0.0:
             return -1e101
 
         return numpy.log(like_total)
+
+    def array_to_scalar(self, var):
+        # numpy no longer accepts ragged arrays,
+        # check that if var is an array it only has one element and return it
+        if hasattr(var, '__len__'):
+            try:
+                assert(len(var) == 1)
+            except AssertionError:
+                raise ValueError('Malformed likelihood')
+            var = var[0]
+        return var
