### 2024-07-02 55963e2
Removed alive_progress as a dependency.
We may want to reverse this since it's pretty nice.
	modified:   environment.yml
	modified:   neost/PosteriorAnalysis.py
