### 2024-07-02 cad85e7
Added python to environment.yml to ensure conda installs a separate python in the environment. Also added some comments in environment.yml to explain the specific version requirements. 	modified:   environment.yml
