+++ docs/source/acknowledgements.rst
@@ -18,12 +18,13 @@ Core team
 
 Those involved in code maintenance and development (technical and scientific).
 
+* Melissa Mendes (mmendes[at]theorie.ikp.physik.tu-darmstadt.de), TU Darmstadt
 * Geert Raaijmakers (g.raaijmakers93[at]gmail.com), the original developer of NEoST
 * Patrick Timmerman (patrick.timmerman98[at]gmail.com)
 * Nathan Rutherford (nathan.rutherford[at]unh.edu), University of New Hampshire
 * Tuomo Salmi (t.h.j.salmi[at]uva.nl), University of Amsterdam
+* Isak Svensson (isak.svensson[at]physik.tu-darmstadt.de), TU Darmstadt
 * Anna Watts (a.l.watts[at]uva.nl), University of Amsterdam
-* Chanda Prescod-Weinstein, University of New Hampshire
 
 Community
 ~~~~~~~~~
@@ -36,6 +37,7 @@ improvements.
 * Kai Hebeler
 * Daniela Huppenkothen
 * Samaya Nissanke
+* Chanda Prescod-Weinstein
 * Achim Schwenk
 
 Graphical design
