### 2024-07-04 e4bd8d7
Changed conda environment name from neost to NEoST 	modified:   environment.yml
