### 2024-07-05 a115394
Added myself and Melissa to acknowledgments 	modified:   docs/source/acknowledgements.rst
