### 2024-07-05 8b2122d
Updated installation instructions 	modified:   docs/source/install.rst
