+++ docs/source/install.rst
@@ -3,8 +3,8 @@
 Installation
 ============
 
-NEoST is a Python software package that performs neutron star equation 
-of state inference process. it is an open source package that was 
+NEoST is a Python software package that performs neutron star equation
+of state inference process. it is an open source package that was
 and is available on `GitHub`_ and can be cloned as:
 
 .. _GitHub: https://github.com/xpsi-group/neost.git
@@ -13,12 +13,12 @@ and is available on `GitHub`_ and can be cloned as:
 
 	git clone https://github.com/xpsi-group/neost.git </path/to/neost>
 
-On this page we will describe how to install NEoST and it's prerequisites 
+On this page we will describe how to install NEoST and it's prerequisites
 on your local machine.
 
 .. note::
 	
-	For installation on a high-performance computing system, we direct 
+	For installation on a high-performance computing system, we direct
 	the reader to the alternative instructions at the bottom of this page.
 
 .. _conda_env:
@@ -26,119 +26,64 @@ on your local machine.
 Prerequisite Python Packages
 ----------------------------
 
-NEoST was developed in Python 3, as such we recommend creating a conda virtual 
-environment with anaconda3 according to the instructions below. This will 
+NEoST was developed in Python 3, as such we recommend creating a conda virtual
+environment with anaconda3 or miniconda3 according to the instructions below. This will
 ensure installing NEoST won't break any existing software packages.
 
 .. _basic_env:
 
-If Python 3 has been installed using the Anaconda Distribution, a new virtual 
-Python environment can then be created by entering the following command 
-into the terminal:
+If Python 3 has been installed using the Anaconda Distribution, a new virtual
+Python environment can then be created by navigating to the NEoST base directory
+and entering the following command into the terminal:
 
 .. code-block:: bash
 
-	conda create --name NEoST --clone base
+	conda env create -f environment.yml
 
-This command will clone the base Python environment into a new environment 
-called NEoST. Next, in order to be able to install packages on the new NEoST 
-environment enter the following: 
+This command will create a new conda environment called NEoST and install all dependencies.
+If you don't want to use conda, you can install the dependencies listed in environment.yml
+in your preferred way.
+
+.. note::
+
+	The latest versions of Cython (3.*) are not currently compatible with NEoST.
+	Consequently, Numpy 2.0 is also incompatible with NEoST as older Cython versions rely on Numpy 1.*.
+
+Next, in order to be able to install NEoST itself in the new environment, enter the following:
 
 .. code-block:: bash
 
 	conda activate NEoST
 
-This command changes the active virtual environment from the default base 
-environment to the new NEoST environment and needs to be entered any time 
+This command changes the active virtual environment from the default base
+environment to the new NEoST environment and needs to be entered any time
 NEoST is used.
 
-A next step is to make sure that a number of packages have a specific
-version of them installed, these are the following packages and versions:
+Installing NEoST
+----------------
+
+With the prerequisites out of the way, NEoST can now be installed. First
+open a new terminal in the directory where NEoST's source code was extracted
+and enter the following command:
 
 .. code-block:: bash
 
-	alive_progress v2.4.1
-	contextlib v0.6.0.post1
-	corner v2.2.1
-	GetDist v1.3.4
-	Matplotlib v3.3.4
-	NumPy v1.21.2
-	SciPy v1.6.2
-	Seaborn v0.11.1
-	
-These packages can be installed with the following commands:
+	python build.py install
+
+Alternatively, NEoST can also be installed without cythonizing the TOV solvers, however this results
[…]
