### 2024-07-05 56ab190
Added new entry in CHANGELOG 	modified:   CHANGELOG.rst
