### 2024-07-05 a0e8d42
Minor edit 	modified:   docs/source/install.rst
