### 2024-07-05 5f68d97
Added Cython 3 and Numpy 2 compatibility in the "Future" section 	modified:   docs/source/todo.rst
