+++ environment.yml
@@ -3,7 +3,7 @@ channels:
         - defaults
         - conda-forge
 dependencies:
-	- python # Just to ensure we're not using system-wide python. See PEP 668
+        - python # Just to ensure we're not using system-wide python. See PEP 668
         - numpy <= 1.26.4 # Required because cython < 3 is incompatible with numpy 2.0
         - cython ~= 0.29 # NEoST is currently incompatible with cython 3
         - matplotlib
