### 2024-07-06 f8f2698
Added "noexcept" to two function declarations in the Cython TOV solvers. This makes them compatible with Cython3.
	modified:   neost/tovsolvers/TOVh.pyx
	modified:   neost/tovsolvers/TOVr.pyx
