### 2024-07-06 748d75c
Renamed build.py to setup.py to enable installing through "pip install ." 	modified:   docs/source/install.rst 	renamed:    build.py -> setup.py
