### 2024-07-06 b28f2f6
setup.py: make the description a single line, since linebreaks are not allowed anymore 	modified:   setup.py
