### 2024-07-06 cdb5fad
Fixed some "invalid escape sequence" warnings in plotting functions 	modified:   neost/PosteriorAnalysis.py 	modified:   neost/Star.py
