### 2024-07-06 022804d
Improved the wording of the "No Cython TOV solvers" message 	modified:   neost/Star.py
