### 2024-07-06 a247d8a
Removed cython and numpy version restrictions in environment.yml since they no longer apply 	modified:   environment.yml
