+++ environment.yml
@@ -4,8 +4,8 @@ channels:
         - conda-forge
 dependencies:
         - python # Just to ensure we're not using system-wide python. See PEP 668
-        - numpy <= 1.26.4 # Required because cython < 3 is incompatible with numpy 2.0
-        - cython ~= 0.29 # NEoST is currently incompatible with cython 3
+        - numpy
+        - cython
         - matplotlib
         - scipy
         - pymultinest
