### 2024-07-06 2757b14
Edited documentation to remove the Cython and Numpy version restrictions 	modified:   docs/source/install.rst 	modified:   docs/source/todo.rst
