+++ CHANGELOG.rst
@@ -44,9 +44,11 @@ Fixed
 ^^^^^
 Compatibility issues:
 * A couple of numpy "ragged arrays" problems, which numpy no longer supports. Two in neost/Likelihood.py, and one in the initial_conditions() function in neost/tovsolvers/TOVr_python.py. The ragged arrays were previously constructed similar to np.array([x1, [x2], x3, [x4]]) whereas in this version they are constructed like np.array([x1, x2, x3, x4]).
+* Cython 3 compilation issue.
 * Use "density" instead of "normed" in numpy.histogramdd.
 * seaborn.kdeplot: Use "fill" instead of "shade", "levels" instead of "n_levels", "cmap" instead of "colors", and modify the supplied values accordingly
 * Function name change: scipy.integrate.cumtrapz is now called scipy.integrate.cumulative_trapezoid. This was introduced in scipy 1.6 and the old name will become deprecated in scipy 1.14.
+* Some minor plotting warnings
 
 Added
 ^^^^^
@@ -65,6 +67,7 @@ Removed
 Changed
 ^^^^^^^
 * Cython TOV solvers no longer print "using c code". Python TOV solvers, on the other hand, now raise a warning when they are in use.
+* Renamed build.py to setup.py to enable installing with pip
 
 Deprecated
 ^^^^^^^^^^
