### 2024-07-08 5d153a5
Updated version number to 0.10.0 	modified:   CHANGELOG.rst 	modified:   docs/source/conf.py 	modified:   neost/__init__.py 	modified:   setup.py
