+++ neost/__init__.py
@@ -1,6 +1,6 @@
 """ NEoST: Open-source code for equation of state inference via nested sampling. """
 from __future__ import print_function
-__version__ = "0.9.0"
+__version__ = "0.10.0"
 __author__ = "Geert Raaijmakers"
 
 try:
