### 2024-07-08 a4a2ff2
Update ci_tests.yml
Change CI tests to reflect renaming of build.py to setup.py
