### 2024-07-08 329e6ce
Added jupyter to environment.yml, with a note that it's only for running the notebooks in docs/source/ 	modified:   environment.yml
