+++ environment.yml
@@ -16,3 +16,4 @@ dependencies:
         - corner
         - gsl
         - kalepy
+        - jupyter # Optional, only required for running the example notebooks in docs/source/
