### 2024-07-08 fce55f1
Addressed Anna's comments on the pull request 	modified:   CHANGELOG.rst 	modified:   docs/source/install.rst 	modified:   neost/__init__.py
