### 2024-07-08 34d173f
Added python 3.12 to the CI tests 	modified:   .github/workflows/ci_tests.yml
