### 2024-07-08 459894e
Fixed the "a3 parameter issue" by renaming "a3" to "a3/a2" 	modified:   neost/eos/speedofsound.py 	modified:   examples/CS_example.py 	modified:   examples/CS_prior_example.py
