### 2024-07-08 f7f4aac
Update acknowledgements.rst
