+++ docs/source/overview.rst
@@ -63,7 +63,7 @@ Crust models
 
 For densities below 0.5 :math:`n_s`, where :math:`n_s = 0.16 \mathrm{fm}^{-3}`,  NEoST uses the BPS EoS for the outer crust (`Baym, Pethick and Sutherland 1971  <https://ui.adsabs.harvard.edu/abs/1971ApJ...170..299B/abstract>`_). 
 
-From this density until a core transition density (typically 1.1 :math:`n_s`) NEoST assumes a polytrope with a varying normalization that captures the range of allowed EoS from chiral effective field theory (`cEFT  <https://ui.adsabs.harvard.edu/abs/2010PhRvC..82a4314H/abstract>`_).  This intermediate part is matched to the models for outer crust and core.  Various cEFT prescriptions are available:  `Hebeler et al. (2013) <https://ui.adsabs.harvard.edu/abs/2013ApJ...773...11H/abstract>`_; `Tews et al. (2013) <https://ui.adsabs.harvard.edu/abs/2013PhRvL.110c2504T/abstract>`_; `Lynn et al. (2016) <https://ui.adsabs.harvard.edu/abs/2016PhRvL.116f2501L/abstract>`_; and `Drischler et al. (2019) <https://ui.adsabs.harvard.edu/abs/2019PhRvL.122d2501D/abstract>`_.
+From this density until a core transition density (typically between 1.1 to 1.5 :math:`n_s`) NEoST assumes a polytrope with a varying normalization that captures the range of allowed EoS from chiral effective field theory (`cEFT  <https://ui.adsabs.harvard.edu/abs/2010PhRvC..82a4314H/abstract>`_). This intermediate part is matched to the parametrized models for core, at higher densities.  Various cEFT prescriptions are available:  `Hebeler et al. (2013) <https://ui.adsabs.harvard.edu/abs/2013ApJ...773...11H/abstract>`_; `Tews et al. (2013) <https://ui.adsabs.harvard.edu/abs/2013PhRvL.110c2504T/abstract>`_; `Lynn et al. (2016) <https://ui.adsabs.harvard.edu/abs/2016PhRvL.116f2501L/abstract>`_; and `Drischler et al. (2019) <https://ui.adsabs.harvard.edu/abs/2019PhRvL.122d2501D/abstract>`_. In the most recent version of NEoST, we have also added `Keller et al. (2023) <https://ui.adsabs.harvard.edu/abs/2023PhRvL.130g2701K/abstract>`_ :math:`N^{2}LO` and :math:`N^{3}LO` cEFT EOS, with calculated error bars up to 1.5 :math:`n_s`.  
 
 Core models
 ^^^^^^^^^^^
