### 2024-07-08 192913d
Added Achim's ERC grant to acknowledgements and sorted contributor list alphabetically 	modified:   docs/source/acknowledgements.rst
