### 2024-07-08 0bc8cd6
Update build_docs.yml
Changed python build.py install to python setup.py install
