### 2024-07-08 83b8783
Removed multinest as an explicit dependency because for some reason it could cause the creation of the conda environment to fail due (probably) to an out-of-memory error. Multinest is still installed as a dependency of pymultinest. 	modified:   environment.yml
