+++ environment.yml
@@ -9,7 +9,6 @@ dependencies:
         - matplotlib
         - scipy
         - pymultinest
-        - multinest
         - pytest
         - numba
         - seaborn
