### 2024-07-08 4b4de0f
Updated copyright year to read "2016-2024" in all documents where it's mentioned 	modified:   LICENSE 	modified:   README.rst 	modified:   docs/source/conf.py
