### 2024-07-08 9cc595d
Updated installation instructions slightly, mainly to reflect the correct environment name ("neost"), but also with better wording and a mention that pandoc is required for building the documentation 	modified:   docs/source/install.rst
