### 2024-07-08 37a3546
Removed pandoc from environment.yml, and added it to build_docs.yml 	modified:   .github/workflows/build_docs.yml 	modified:   environment.yml
