+++ environment.yml
@@ -16,4 +16,3 @@ dependencies:
         - gsl
         - kalepy
         - jupyter # Optional, only required for running the example notebooks in docs/source/
-        - pandoc
