+++ .github/workflows/build_docs.yml
@@ -34,7 +34,7 @@ jobs:
     - name : install docs dependencies
       run: |
         micromamba install sphinx decorator h5py
-        micromamba install -c conda-forge nbsphinx nbconvert
+        micromamba install -c conda-forge nbsphinx nbconvert pandoc
         pip install sphinxcontrib-websupport sphinx_rtd_theme
     - name: install neost package
       run: |
