### 2024-07-08 df1ea34
Removed some unnecessary stuff from the README 	modified:   README.rst
