+++ README.rst
@@ -33,8 +33,7 @@ For more details on current and planned capabilities, check out the
 Installation and Testing
 ------------------------
 
-NEoST has a complex set of dependencies, and is therefore currently best 
-installed from source. The documentation provides
+NEoST is best installed from source. The documentation provides
 `step-by-step installation instructions <https://xpsi-group.github.io/neost/install.html>`_
 for Linux and for limited MacOS systems.
 
