### 2024-07-09 7d1503f
Fixed trivial mistake 	modified:   CHANGELOG.rst
