+++ CHANGELOG.rst
@@ -34,7 +34,7 @@ and this project adheres to
 .. ^^^^^^^^^^^
 
 [v0.10.0 - 2024-07-10]
-~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~
 
 Summary
 ^^^^^^^
