### 2024-07-09 5cf788f
Changed "python setup.py install" to "pip install ." in the CI tests. The former is deprecated. 	modified:   .github/workflows/ci_tests.yml
