+++ .github/workflows/ci_tests.yml
@@ -43,7 +43,7 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: install neost package
       run: |
-        python setup.py install
+        pip install .
       shell: bash -l {0}
 
     - name: Test with pytest
