### 2024-07-09 6549bd9
Fixed minor layout issue 	modified:   CHANGELOG.rst
