### 2024-07-09 5e20fff
Changed beginning copyright year to 2020 	modified:   LICENSE 	modified:   README.rst 	modified:   docs/source/conf.py
