### 2024-07-09 2c10332
Minor readme edit 	modified:   README.rst
