### 2024-07-16 f0640dd
Fixed a numpy "ragged arrays" issue in PosteriorAnalysis.compute_table_data() 	modified:   neost/PosteriorAnalysis.py
