### 2024-07-17 dcee10d
Update PosteriorAnalysis.py
Fixes (ideally) the issues of zeros appearing and only have one entry in table_data.txt
