### 2024-07-17 3002dff
Add new paper to publication list
