### 2024-07-17 804360c
Update PosteriorAnalysis.py
switching i to k in the for loop for the row list
