### 2024-07-18 0b8501b
Update PosteriorAnalysis.py
Deleted printing messages since everything is working properly.
