+++ _sources/install.rst.txt
@@ -1,0 +1,185 @@
+.. _install:
+
+Installation
+============
+
+NEoST is a Python software package that performs neutron star equation
+of state inference process. It is an open source package that is available
+on `GitHub`_ and can be cloned as:
+
+.. _GitHub: https://github.com/xpsi-group/neost.git
+
+.. code-block:: bash
+
+	git clone https://github.com/xpsi-group/neost.git </path/to/neost>
+
+On this page we will describe how to install NEoST and its prerequisites.
+
+Prerequisite Python Packages
+----------------------------
+
+NEoST was developed in Python 3, as such we recommend creating a conda virtual
+environment with anaconda3 or miniconda3 according to the instructions below. This will
+ensure installing NEoST won't break any existing software packages.
+
+.. _basic_env:
+
+Assuming a conda base environment has been installed and activated (see conda
+installation instructions), a new virtual Python environment can then be created
+by navigating to the NEoST base directory and entering the following command into the terminal:
+
+.. code-block:: bash
+
+	conda env create -f environment.yml
+
+This command will create a new conda environment called neost and install all dependencies.
+If you don't want to use conda, you can install the dependencies listed in environment.yml
+in your preferred way.
+
+Next, in order to be able to install NEoST itself in the new environment, enter the following:
+
+.. code-block:: bash
+
+	conda activate neost
+
+This command changes the active virtual environment from the default base
+environment to the new neost environment and needs to be entered any time
+NEoST is used.
+
+Installing NEoST
+----------------
+
+With the prerequisites out of the way, NEoST can now be installed. First (if
+you haven't done so already) navigate to the directory where NEoST's source
+code was extracted and activate the conda environment. Then install NEoST with
+
+.. code-block:: bash
+
+	pip install .
+
+Alternatively, NEoST can be installed without cythonizing the TOV solvers, however this results
+in much slower performance. To do this, enter the following command:
+
+.. code-block:: bash
+
+	python setup.py install --nocython
+
+
+.. Alternative instructions for prerequisites
+.. ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. There are alternative ways to install GSL, MultiNest and PyMultiNest.
+.. To manually install GSL enter the following
+.. into a terminal:
+
+.. code-block .. :: bash
+
+	wget -v http://mirror.koddos.net/gnu/gsl/gsl-latest.tar.gz
+	tar -zxvf gsl-latest.tar.gz
+	cd gsl-latest
+	./configure CC=gcc --prefix=$HOME/gsl
+	make
+	make check
+	make install
+	make installcheck
+	make clean
+
+.. With this done GSL will have to be added to your path, this is done with the
+.. following command:
+
+.. code-block .. :: bash
+
+	export PATH=$HOME/gsl/bin:$PATH
+
+.. This command must be given any time GSL is used, therefore it is recommended
+.. to add this command to your ``~.bashrc`` file.
+
+.. In order to manually install MultiNest and PyMultinest, first install the
+.. prerequisites. These are mpi4py and compilers for c and fortran and can be
+.. installed with the following commands:
+
+.. code-block .. :: bash
+
+	conda install -c conda-forge mpi4py
+	sudo apt-get install cmake libblad-dev liblapack-dev libatlas-base-dev
+
+.. When these have finished installing, clone the MultiNest repository, navigate
+.. to the cloned repository and install MultiNest using the following commands:
+
+.. code-block .. :: bash
+
+	git clone https://github.com/farhanferoz/MultiNest.git <path/to/clone>/multinest
[…]
