+++ _sources/todo.rst.txt
@@ -1,0 +1,11 @@
+.. _TODO:
+
+Future
+------
+
+If there is a feature you would like to see that is not currently
+implemented in the package please let us know!  We are currently working on the following upgrades:
+
+* Adding the facility to use information from EM counterparts of GW events (as done in `Raaijmakers et al. 2021 <https://ui.adsabs.harvard.edu/abs/2021ApJ...918L..29R/abstract>`_).
+* Adding the facility to do inference including dark matter components (as done in `Rutherford et al. 2023 <https://ui.adsabs.harvard.edu/abs/2023PhRvD.107j3051R/abstract>`_).
+* Linking to alternative samplers.
