### 2024-07-26 294df8c
WIP: Mainly use pyproject.toml to define the installation procedure 	new file:   pyproject.toml 	modified:   setup.py
