+++ Makefile
@@ -1,0 +1,6 @@
+.PHONY: build
+build: 
+	pip install .
+
+clean:
+	rm -r ./build/ ./dist/ ./NEoST.egg-info/
