### 2024-08-02 57dc34d
Added kalepy as a runtime dependency 	modified:   pyproject.toml
