### 2024-08-02 09d8fdb
Updated installation instructions with conda-less instructions 	modified:   docs/source/install.rst
