+++ docs/source/install.rst
@@ -1,5 +1,6 @@
 .. _install:
 
+============
 Installation
 ============
 
@@ -11,58 +12,106 @@ on `GitHub`_ and can be cloned as:
 
 .. code-block:: bash
 
-	git clone https://github.com/xpsi-group/neost.git </path/to/neost>
+	git clone https://github.com/xpsi-group/neost.git
 
 On this page we will describe how to install NEoST and its prerequisites.
 
-Prerequisite Python Packages
-----------------------------
+Creating an environment and installing NEoST's dependencies
+===========================================================
 
-NEoST was developed in Python 3, as such we recommend creating a conda virtual
-environment with anaconda3 or miniconda3 according to the instructions below. This will
-ensure installing NEoST won't break any existing software packages.
+NEoST was developed in Python 3. We recommend installing NEoST and its dependencies in a virtual
+environment. There are two main options for creating an environment and installing the dependencies:
+Python's first-party package installer, pip, and the third-party Conda package manager.
+
+You need pip in both cases. It is likely already installed. If not, follow the official instructions for installing it.
+
+Using pip
+---------
+
+Two of NEoST's dependencies, GSL (GNU Scientific Library) and MultiNest, are not available in the Python package index (PyPI),
+and need to be installed separately.
+
+GSL is widely available and can almost always be installed with your operating system's package manager,
+or loaded with something like ``module load gsl`` on a computing cluster.
+Alternatively you can compile it yourself using the official instructions.
+
+However you choose to obtain GSL, make sure it works by executing
+
+.. code-block:: bash
+
+	gsl-config --version
+
+which should print the installed GSL version.
+
+MultiNest, on the other hand, must likely be compiled manually. Follow these steps (which assume that you have ``git``, a ``C++`` compiler, ``cmake``, and ``make`` installed):
+
+.. code-block:: bash
+
+	git clone https://github.com/JohannesBuchner/MultiNest
+	cd MultiNest/build
+	cmake ..
+	make
+
+You then need to add ``<path/to>/MultiNest/lib`` to an environment variable called ``LD_LIBRARY_PATH`` using
+
+.. code-block:: bash
+
+	export LD_LIBRARY_PATH=<path/to>/MultiNest/lib:$LD_LIBRARY_PATH
+
+Of course, exchange ``<path/to>`` for the actual path where you cloned MultiNest. You can add the export command
+to your ``$HOME/.bashrc`` file if you don't want to execute it in every new terminal you're running NEoST in.
+
+The next step is to create a virtual environment and install NEoST. In Python, the official way to create
+a virtual environment is by using the venv module:
+
+.. code-block:: bash
+
+	python -m venv path/to/environment
+
+where you should substitute path/to/environment for a location of your choice. Then activate the environment with
+
+.. code-block:: bash
+
+	source path/to/environment/bin/activate
+
+Using Conda
+-----------
 
 .. _basic_env:
 
 Assuming a conda base environment has been installed and activated (see conda
-installation instructions), a new virtual Python environment can then be created
-by navigating to the NEoST base directory and entering the following command into the terminal:
+installation instructions), a new virtual environment can then be created
+by navigating to the NEoST base directory and executing
 
 .. code-block:: bash
 
 	conda env create -f environment.yml
 
-This command will create a new conda environment called neost and install all dependencies.
-If you don't want to use conda, you can install the dependencies listed in environment.yml
-in your preferred way.
-
-Next, in order to be able to install NEoST itself in the new environment, enter the following:
+This will create a new conda environment called neost and install all dependencies (including GSL and MultiNest).
+Once the environment has been created, activate it with
 
 .. code-block:: bash
 
 	conda activate neost
 
-This command changes the active virtual environment from the default base
[…]
