### 2024-08-02 4f614e9
Changed non-functional "python setup.py install" to "pip install ." 	modified:   .github/workflows/build_docs.yml
