### 2024-08-02 cbd2416
Added jupyter as a dependency. It's of course only needed for running jupyter notebooks of course, but we have it in environment.yml so it should be in pyproject.toml too. 	modified:   pyproject.toml
