### 2024-08-02 c860d44
Reinstated tested instructions for building the documentation. 	modified:   docs/source/install.rst
