### 2024-08-02 98d2b41
Improved installation instructions 	modified:   docs/source/install.rst
