### 2024-08-02 4cf8ecb
Improved Makefile and mention it in the installation instructions 	modified:   Makefile 	modified:   docs/source/install.rst
