### 2024-08-02 39fb90b
Removed commented-out stuff and leftovers from jupyter notebook conversion 	modified:   examples/Tabulated_example.py 	modified:   examples/Tabulated_prior_example.py
