+++ examples/Tabulated_prior_example.py
@@ -1,11 +1,5 @@
-#!/usr/bin/env python
-# coding: utf-8
-
 # # Tabulated Prior Example
 # The machinary and the explanation of them is identical to those found in the Piecewise Polytropic, Speed of Sound, and Tabulated Examples.
-
-# In[1]:
-
 
 import neost
 from neost.eos import tabulated
@@ -34,10 +28,6 @@ rho_ns = global_imports._rhons
 
 # Define name for run
 run_name = "prior-tabulated-"
-
-
-# In[ ]:
-
 
 # We're exploring a tabulated (T) EoS 
 eos_name = 'tabulated'
@@ -85,10 +75,6 @@ variable_params = {}
 for i in range(number_stars):
 	variable_params.update({'rhoc_' + str(i+1):[14.6, 16]})
 
-
-# In[ ]:
-
-
 # Define static parameters, empty dict because all params are variable 
 static_params={}
 
@@ -103,22 +89,6 @@ print("number of parameters is %d" %len(variable_params))
 def loglike(pseudo_var):
     return 1.
 
-
-# In[ ]:
-
-
-# No prior & likelihood test, there is no likelihood after all
-# print("Testing prior and likelihood")
-# cube = np.random.rand(50, len(variable_params))
-# for i in range(len(cube)):
-#     par = prior.inverse_sample(cube[i])
-#     print(likelihood.call(par))
-# print("Testing done")
-
-
-# In[ ]:
-
-
 # Then we start the sampling, note the greatly increased number of livepoints, this is required because each livepoint terminates after 1 iteration
 start = time.time()
 result = solve(LogLikelihood=loglike, Prior=prior.inverse_sample, n_live_points=100000, evidence_tolerance=0.1,
@@ -126,14 +96,9 @@ result = solve(LogLikelihood=loglike, Prior=prior.inverse_sample, n_live_points=
 end = time.time()
 print(end - start)
 
-
-# In[ ]:
-
-
 # Compute auxiliary data for posterior analysis
 PosteriorAnalysis.compute_auxiliary_data('chains/' + run_name, tabulated_example,
                                          variable_params, static_params, chirp_mass)
-
 
 # Make some analysis plots
 PosteriorAnalysis.cornerplot('chains/' + run_name, variable_params)
