### 2024-08-02 34a46f9
Changed "numpy" to "np" and "pyplot" to "plt". Fixes Issue #33 	modified:   neost/Likelihood.py 	modified:   neost/PosteriorAnalysis.py 	modified:   neost/Prior.py 	modified:   neost/Star.py 	modified:   neost/eos/base.py 	modified:   neost/utils.py
