+++ neost/PosteriorAnalysis.py
@@ -1,9 +1,9 @@
-import numpy
+import numpy as np
 import matplotlib
 from matplotlib.lines import Line2D
 import matplotlib.patches as mpatches
 from matplotlib.colors import ListedColormap
-from matplotlib import pyplot
+import matplotlib.pyplot as plt
 import seaborn as sns
 from scipy.interpolate import UnivariateSpline
 from scipy.stats import gaussian_kde
@@ -22,7 +22,7 @@ Msun = global_imports._M_s
 pi = global_imports._pi
 rho_ns = global_imports._rhons
 
-colors = numpy.array(["#c6878f", "#b79d94", "#969696", "#67697c", "#233b57", "#BCBEC7"])
+colors = np.array(["#c6878f", "#b79d94", "#969696", "#67697c", "#233b57", "#BCBEC7"])
 
 dyncm2_to_MeVfm3 = 1./(1.6022e33)
 gcm3_to_MeVfm3 = 1./(1.7827e12)
@@ -31,13 +31,13 @@ oneoverfm_MeV = 197.33
 
 def m1(mc, m2):
     num1 = (2./3.)**(1./3.)*mc**5.
-    denom1 = (9*m2**7. *mc**5. + numpy.sqrt(3.)*numpy.sqrt(abs(27*m2**14.*mc**10.-4.*m2**9.*mc**15.)))**(1./3.)
+    denom1 = (9*m2**7. *mc**5. + np.sqrt(3.)*np.sqrt(abs(27*m2**14.*mc**10.-4.*m2**9.*mc**15.)))**(1./3.)
     denom2 = 2.**(1./3.)*3.**(2./3.)*m2**3.
     return num1/denom1 + denom1/denom2
 
 def calc_bands(x, y):
-    miny = numpy.zeros((len(y),3))
-    maxy = numpy.zeros((len(y),3))
+    miny = np.zeros((len(y),3))
+    maxy = np.zeros((len(y),3))
     
     for i in range(len(y)):
         z = y[i][y[i]>0.0]
@@ -45,34 +45,34 @@ def calc_bands(x, y):
             print('sample too small for %.2f' %x[i])
             continue
         kde = gaussian_kde(z)
-        testz = numpy.linspace(min(z),max(z), 1000)
+        testz = np.linspace(min(z),max(z), 1000)
         pdf = kde.pdf(testz)
         array = pdf
-        index_68 = numpy.where(numpy.cumsum(numpy.sort(array)[::-1]) < sum(array)*0.6827)[0]
-        index_68 = numpy.argsort(array)[::-1][index_68]
-        index_95 = numpy.where(numpy.cumsum(numpy.sort(array)[::-1]) < sum(array)*0.95)[0]
-        index_95 = numpy.argsort(array)[::-1][index_95]
+        index_68 = np.where(np.cumsum(np.sort(array)[::-1]) < sum(array)*0.6827)[0]
+        index_68 = np.argsort(array)[::-1][index_68]
+        index_95 = np.where(np.cumsum(np.sort(array)[::-1]) < sum(array)*0.95)[0]
+        index_95 = np.argsort(array)[::-1][index_95]
         miny[i] =  x[i], min(testz[index_68]), min(testz[index_95])
         maxy[i] =  x[i], max(testz[index_68]), max(testz[index_95])
         
-    miny = miny[~numpy.all(miny == 0, axis=1)]
-    maxy = maxy[~numpy.all(maxy == 0, axis=1)]
+    miny = miny[~np.all(miny == 0, axis=1)]
+    maxy = maxy[~np.all(maxy == 0, axis=1)]
     return miny, maxy
 
 def get_quantiles(array, quantiles=[0.025, 0.5, 0.975]):
[…]
