+++ examples/Tabulated_prior_example.py
@@ -10,13 +10,9 @@ from neost import PosteriorAnalysis
 from scipy.stats import multivariate_normal
 from scipy.stats import gaussian_kde
 import numpy as np
-import matplotlib
-from matplotlib import pyplot
 from pymultinest.solve import solve
 import time
 import os
-#if not os.path.exists("chains"): 
-#   os.mkdir("chains")
 import neost.global_imports as global_imports
 
 # Some physical constants
