+++ examples/Tabulated_example.py
@@ -7,9 +7,6 @@ from scipy.stats import multivariate_normal
 from scipy.stats import gaussian_kde
 from neost import PosteriorAnalysis
 import numpy as np
-import random
-import matplotlib
-from matplotlib import pyplot
 from pymultinest.solve import solve
 import time
 import os
