### 2024-08-06 #6003 @drannawatts
A couple of updates from our side! We had been planning to hold off on v1.0 release until the JOSS review was complete but since this seems to be taking a bit longer we will likely go ahead and do this since we've made quite a few updates associated with a new NICER EoS publication so it's a good moment to do this.

With that in mind, @adonath, would there be any objection to adding a couple of additional authors to the JOSS paper submission?   We have some folks from TU Darmstadt who have contributed a lot of the code in the most recent updates and it would be nice to acknowledge that if we still have the chance.   We can also update to include a reference to the new NICER EoS paper.

@chaitanyaafle I see that you have one box still unchecked in your review checklist - have you run into issues with the software there that we need to address? If so let us know and we'll try to fix them ASAP!
