### 2024-08-06 #6003 @adonath
Thanks @drannawatts for the update! There is no objection to add more authors, as long as they made a substantial contribution to the package.

We had a bit of bad luck with the second reviewers, I'm in the process of finding one, once again. If the search is not successful within a reasonable timeline of ~4 weeks I will assign myself.
