### 2024-08-06 #6003 @drannawatts
Thanks @adonath - we'll update the JOSS manuscript within the next few days once we have the final publication details for the new papers.  And no worries on the timeline, as an Editor for MNRAS I know how tricky finding referees can be, and JOSS is asking for a lot more work than a standard paper review!!
