### 2024-09-05 08e47e0
Updating Joss Paper
Anna's newest version of the JOSS paper with citations updates and including Isak and Melissa as co-authors
