### 2024-09-11 990264a
Added v1.0.0 changelog
	modified:   CHANGELOG.rst
