### 2024-09-11 #6003 @nr1118
Hi @adonath, we have updated the NEoST JOSS manuscript and tagged v1.0 as mentioned by @drannawatts. I am also going to run the editorialbot below to generate the new pdf for convenience. Please let us know if you have any questions :)
