### 2024-09-11 b57f160
Update conf.py
Changed tag from 0.10.0 to 1.0
