### 2024-09-11 1833806
Update pyproject.toml
Updated version number to 1.0
