+++ docs/source/conf.py
@@ -77,9 +77,9 @@ author = 'NEoST Core Team'
 # built documents.
 #
 # The short X.Y version.
-version = '1.0'
+version = '1.0.0'
 # The full version, including alpha/beta/rc tags.
-release = '1.0'
+release = '1.0.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
