+++ docs/source/conf.py
@@ -77,7 +77,7 @@ author = 'NEoST Core Team'
 # built documents.
 #
 # The short X.Y version.
-version = '1.0.0'
+version = '1.0'
 # The full version, including alpha/beta/rc tags.
 release = '1.0.0'
 
