+++ joss/neostjoss.bib
@@ -131,7 +131,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Rutherford:2024,
-       author = {{Rutherford}, N. and {Mendes}, M. and {Svensson}, I. and {Schwenk}, A. and {Watts}, A. L. and {Hebeler}, K. and {Keller}, J. and {Prescod-Weinstein}, C. and {Choudhury}, D. and {Raaijmakers}, G. and {Salmi}, T.and {Timmerman}, P. and {Vinciguerra}, S. and {Guillot}, S. and {Lattimer}, J. M.},
+       author = {{Rutherford}, Nathan and {Mendes}, Melissa and {Svensson}, Isak and {Schwenk}, Achim and {Watts}, Anna L. and {Hebeler}, Kai and {Keller}, Jonas and {Prescod-Weinstein}, Chanda and {Choudhury}, Devarshi and {Raaijmakers}, Geert and {Salmi}, Tuomo and {Timmerman}, Patrick and {Vinciguerra}, Serena and {Guillot}, Sebastien and {Lattimer}, James M.},
         title = "{Constraining the Dense Matter Equation of State with New NICER Mass-Radius Measurements and New Chiral Effective Field Theory Inputs}",
       journal = {Astrophysical Journal, Letters},
      keywords = {Gravitational wave sources, Neutron stars, Neutron star cores, X-ray sources, Nuclear astrophysics, 677, 1108, 1107, 1822, 1129, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics, Nuclear Experiment, Nuclear Theory},
