### 2024-09-12 #6003 @nr1118
I apologize for any confusion this may cause (found a few typos in the manuscript), but please use the generated pdf of the JOSS manuscript below this message.
