### 2024-09-30 e1069c3
Including changes for dark matter
-updated CHANGELOG with dark matter update changes/additions
-updated setup.py to include TOVdm.pyx solver and to print NEoST-v2.0.0 messaging during install
