+++ setup.py
@@ -1,77 +1,167 @@
-# This file specifies how the Cython TOV solvers are to be compiled.
-# It is not necessary at all if only Python TOV solvers are to be used.
+"""
+To compile to C code, and then compile the C code and link libraries, and
+install as a Python package:
+    --> python setup.py install [--user]
+To build extensions in place and not install:
+    --> python setup.py build_ext -i
+The extension is then constructed in the source directory. To import NEoST,
+ensure the sys.path searches the ``src`` directory. If a Python module is
+modified, the package then does not need to be reinstalled for usage, but for
+use in an interactive environment, the module must be reloaded
+(or the kernel must be restarted).
+"""
 
-# Standard libraries
-from setuptools import Extension, setup
-import subprocess as sub
-import sys
+import os
 
-# 3rd party
-import numpy as np
-
-OS = sys.platform
-
-# Check the operating system
-if 'darwin' not in OS and 'linux' not in OS:
-    raise ValueError('Unsupported operating system ' + OS)
-
-# Make sure GSL is installed
 try:
-    gsl_version = sub.check_output(['gsl-config','--version'])[:-1].decode("utf-8")
-    gsl_prefix = sub.check_output(['gsl-config','--prefix'])[:-1].decode("utf-8")
-except Exception:
-    print('GNU Scientific Library cannot be located.')
-    raise
-else:
-    print('GSL version: ' + gsl_version)
-    libraries = ['gsl','gslcblas','m']
-    library_dirs = [gsl_prefix + '/lib']
+    from setuptools import setup, find_packages, Extension
+except ImportError:
+    from distutils.core import setup, find_packages, Extension
 
-# Common includes, linker arguments, compiler arguments
-include_dirs = [np.get_include(), gsl_prefix+'/include', '.']
-extra_link_args = ['-fopenmp']
-extra_compile_args = ['-fopenmp', '-Wno-unused-function', '-Wno-uninitialized']
+if __name__ == '__main__':
+    import numpy
+    import sys
+    OS = sys.platform
+    print(sys.path)
 
-# OS-specific settings
-if 'darwin' in OS:
-    # Using compiler of clang with llvm installed
-    os.environ["CC"] = "/usr/local/opt/llvm/bin/clang"
-    os.environ["CXX"] = "/usr/local/opt/llvm/bin/clang++"
-    library_dirs.append('/usr/local/opt/llvm/lib')
-    library_dirs.append('/opt/local/lib')
-    extra_compile_args.append('-Wno-#warnings')
-    extra_compile_args.append('-Wno-error=format-security')
-    include_dirs.append(['/usr/local/include', '/usr/local/opt/llvm/include', './neost'])
+    if '--nocython' in sys.argv:
+        sys.argv.remove('--nocython')
+        setup(
+            name='NEoST',
+            version='2.0.0',
+            author='NEoST core team',
+            author_email='A.L.Watts@uva.nl',
+            url='',
+            license='GNU GPLv3',
+            description="""NEoST: An open-source code for dense matter equation of state inference via nested sampling.""",
+            packages=find_packages(),
+            install_requires=[],
+            include_package_data=True,
+            classifiers=['Intended Audience :: Science/Research',
+                         'Operating System :: Mac OS X, Linux',
+                         'License :: OSI Approved :: GNU GPLv3 License',
+                         'Programming Language :: Python'])
 
-else:
-    # point to shared library at compile time so runtime resolution
-    # is not affected by environment variables, but is determined
-    # by the binary itself
-    extra_link_args.append('-Wl,-rpath=%s' % (gsl_prefix + '/lib'))
-    extra_compile_args.append('-Wno-cpp')
[…]
