### 2024-09-30 cbe4985
Including examples with dark matter
-added three new examples which show how to compute the MR and MLambda relations within our ADM framework (ADM_MR_Tidal_tuorial), how to include ADM in a prior script (PP_with_ADM_prior_example), and how to include ADM in a posterior inference script (PP_with_ADM_example).
