### 2024-09-30 c1fe989
Updates to documentation with dark matter included
-compiled example notebooks (ADM_MR_Tidal_tutorial, PP_with_ADM_example, and PP_with_ADM_prior_example)
-mathematical description of including ADM in our inference framework
-documentation on new ADM functions within the code.
-
