### 2024-09-30 6ffd069
Dark matter updates to the eos files
Base.py includes:
-a find_epsdm_cent function which computes the ADM central energy density corresponding to a given ADM fraction and baryonic central energy density. Then, solves it as  a root finding problem. 
-a fchi_calc function, which computes the ADM mass-fraction for a given ADM central energy density and baryonic central energy density.

Polytropes, speedofsound, tabulated includes:
-Initialization of the Bosonic or Fermionic ADM equations of state using the Nelson et al 2018 ADM model.
