+++ neost/eos/speedofsound.py
@@ -36,7 +36,13 @@ class SpeedofSoundEoS(BaseEoS):
         If True a low-density cEFT parameterization is used.
     ceft_method: str
         The name of the cEFT calculations used at low density.
-        Can be one of 'Hebeler', 'Drischler', 'Lynn' or 'Tews'.
+        Can be one of 'Hebeler', 'Drischler', 'Lynn', 'Keller-N2LO', 'Keller-N3L0', or 'Tews'.
+    adm_type: str
+        The name of the ADM particle type. Can be 'None', 'Bosonic', or 'Fermionic'
+    dm_halo: bool
+        If True, ADM halos will be allowed.
+    two_fluid_tidal: bool
+        If True, the two-fluid tidal deformability solver will be used. See TidalDef.py file
 
     Methods
     -------
@@ -46,6 +52,8 @@ class SpeedofSoundEoS(BaseEoS):
         Construct the crust of the equation of state, with or without cEFT.
     get_eos()
         Construct the high-density parameterization of the equation of state.
+    add_adm_eos()
+        Function to compute the ADM EoS, whether it be bosonic or fermionic in nature.
     plot()
         Plot the equation of state.
     plot_massradius()
@@ -53,7 +61,7 @@ class SpeedofSoundEoS(BaseEoS):
 
     """
 
-    def __init__(self, crust, rho_t):
+    def __init__(self, crust, rho_t,adm_type = 'None',dm_halo = False,two_fluid_tidal = False):
 
         super(SpeedofSoundEoS, self).__init__(crust, rho_t)
 
@@ -61,6 +69,16 @@ class SpeedofSoundEoS(BaseEoS):
         self.param_names = ['a1', 'a2', 'a3/a2', 'a4', 'a5']
         if self.ceft is True:
             self.param_names.append('ceft')
+
+        self.adm_type = adm_type
+        self.dm_halo = dm_halo
+        self.two_fluid_tidal = two_fluid_tidal
+
+        if self.adm_type not in ['Bosonic', 'Fermionic', 'None']:
+            raise TypeError('ADM model not recognized. Choose either "Bosonic" or "Fermionic". If ADM is not needed, select "None".')
+        
+        if self.adm_type in ['Bosonic','Fermionic']:
+            self.param_names +=['mchi','gchi_over_mphi', 'adm_fraction']
 
 
     def get_eos(self):
@@ -91,7 +109,8 @@ class SpeedofSoundEoS(BaseEoS):
 
         self.eos = UnivariateSpline(self.energydensities,
                                     self.pressures, k=1, s=0)
-
+        if self.adm_type in ['Bosonic', 'Fermionic']:
+            self.add_adm_eos()
 
     #######################
     # Auxiliary functions #
@@ -123,6 +142,98 @@ class SpeedofSoundEoS(BaseEoS):
         return (dmin * self._cs_crust(x * rho_ns) / c**2. +
                 dplus * (self.Cs_model_core(x, norm)))
 
+
+    def add_adm_eos(self):
+
+        def joinarray(ar1,ar2):
+            return np.concatenate((ar1,ar2))
+        
+        _c = 2.99792458e8  # m/s
+        _hbar = 1.0545718e-34  # J*s
+        MeV_to_Joules = 1.60218e-13  # J
+
+        self.mchi = self.eos_params['mchi'] * MeV_to_Joules / _c**2.
+        self.gchi_over_mphi = (self.eos_params['gchi_over_mphi'] / MeV_to_Joules * _c**2.)
+        self.adm_fraction = self.eos_params['adm_fraction']
+
+
+
+        # Bosonic Dark matter component
+        # geometric
+        if self.adm_type == 'Bosonic':
+            #print('Using bosonic EoS')
+            pos_nchi = np.logspace(-5, 55, 5000)
[…]

[…]
