### 2024-09-30 be9a522
Update TidalDef.py
Dropped comments
