### 2024-09-30 87bf2c0
Update TOVdm_python.py
Dropped comments that are not useful
