### 2024-09-30 6680ecc
Update Likelihood.py
Dropped comments that are not useful
