### 2024-09-30 b8315b3
Update Star.py
