### 2024-09-30 78e520e
Update ADM_MR_Tidal_tutorial.py
