+++ neost/tovsolvers/TOVdm.pyx
@@ -1,0 +1,419 @@
+# cython: cdivision=True
+# cython: boundscheck=False
+# cython: nonecheck=False
+# cython: wraparound=False
+
+from libc.math cimport sqrt, sin, cos, acos, log10, pow, exp, pi, log, floor
+from libc.stdio cimport printf, setbuf, stdout
+from libc.stdlib cimport malloc, free, abs
+
+import numpy as np
+
+from GSL cimport *
+from .. import global_imports
+from cython.operator import dereference
+from matplotlib import pyplot
+import time
+cimport numpy as np
+
+
+cdef double c = global_imports._c
+cdef double G = global_imports._G
+cdef double Msun = global_imports._M_s
+
+
+
+cdef int binarySearch(double arr[], int low, int high, double key) nogil:
+    cdef int mid 
+    while high > low:
+        mid = (low + high) / 2
+        if low > high:
+            break
+        if arr[mid] > key:
+            high = mid
+        else:
+            low = mid + 1
+    return low
+
+
+# in terms of r
+cdef double EofP(double pressure, double eps[], double pres[], int idx) nogil:
+    if idx == 0:
+        eds = eps[0] * pow(pressure / pres[0], 3. / 5.)
+    if idx > 0:
+        ci = log(pres[idx] / pres[idx - 1]) / log(eps[idx] / eps[idx - 1])
+        eds = eps[idx - 1] * pow(pressure / pres[idx - 1], 1. / ci)
+    return eds
+
+cdef double PofE(double epsilon, double pres[], double eps[], int idx) nogil:
+    if idx==0:
+        pressure = pres[0]*pow(epsilon/eps[0], 5./3.)
+    if idx>0.:
+        ci = log(pres[idx]/pres[idx-1])/log(eps[idx]/eps[idx-1])
+        pressure = pres[idx-1] * pow(epsilon/eps[idx-1], ci)
+    return pressure
+
+
+
+cdef int TOV_single(double r, const double y[], double f[], void * par) noexcept nogil:
+
+    cdef double p
+    cdef double eps
+
+    cdef double *rhotest = (<double**> par)[0]
+    cdef double *prestest = (<double**> par)[1]
+    cdef double *num_double = (<double**>par)[2]
+    cdef int num = <int> num_double[0]
+    p = sqrt(y[0]*y[0])
+    cdef int idx = binarySearch(prestest, 0, num, p)
+    eps = EofP(p, rhotest, prestest, idx)
+
+
+    f[0] = -(eps + p) * (y[1] + 4. * pi * pow(r,3) * p)
+    f[0] *= pow(r * (r - 2.*y[1]), -1)
+
+
+    f[1] = 4.*pi*pow(r,2) * eps
+
+
+    f[2] = (y[1] + 4.*pi*pow(r,3) * p) * pow(r*(r - 2.*y[1]), -1)
+
+    return GSL_SUCCESS
+
+
+cdef int TOV_complete(double r, const double y[], double f[],void * par) noexcept nogil:
+
+    cdef double pb
+    cdef double epsb
+
+    cdef double pdm
+    cdef double epsdm
+
+    cdef double *epsb_test = (<double**> par)[0]
+    cdef double *presb_test = (<double**> par)[2]
+
+    cdef double *epsdm_test = (<double**> par)[1]
+    cdef double *presdm_test = (<double**> par)[3]
+
+    cdef double *numb_double = (<double**> par)[4]
+    cdef int numb = <int> numb_double[0]
+
+    cdef double *numdm_double = (<double**> par)[5]
[…]
