### 2024-10-01 a42e85c
Update ADM_MR_Tidal_tutorial.py
minor fixes to use plt over pyplot.
