### 2024-10-01 11ac57a
Update CHANGELOG.rst
