+++ .github/workflows/ci_tests.yml
@@ -24,7 +24,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install Conda environment with Micromamba
-      uses: mamba-org/setup-micromamba
+      uses: mamba-org/setup-micromamba@v1
       with:
         environment-file: environment.yml
         cache-env: true
