### 2024-10-07 19cd909
updates to install
