### 2024-10-07 ac7ec78
Updates to install on MAC M1 architectures
