### 2024-10-07 ff5a3cf
Update setup.py
seems like we forgot to include import os
