+++ setup.py
@@ -5,7 +5,7 @@
 from setuptools import Extension, setup
 import subprocess as sub
 import sys
-
+import os
 # 3rd party
 import numpy as np
 
