### 2024-10-07 f3f79fc
updated micromamba install environment for tests
