### 2024-10-07 1b50ea2
Added missing "import os" statement, needed for Mac users 	modified:   setup.py
