### 2024-10-07 2f5144c
Removed redundant "import os" (it got added twice) 	modified:   setup.py
