### 2024-10-08 6c17e7c
Changed "pip install ." to the safer "python3 -m pip install ." 	modified:   Makefile
