+++ _sources/install.rst.txt
@@ -1,0 +1,214 @@
+.. _install:
+
+============
+Installation
+============
+
+NEoST is a Python software package that performs neutron star equation of state inference process. It is an open source package that is available on `GitHub`_ and can be cloned as:
+
+.. _GitHub: https://github.com/xpsi-group/neost.git
+
+.. code-block:: bash
+
+	git clone https://github.com/xpsi-group/neost.git
+
+On this page we will describe how to install NEoST and its prerequisites.
+
+Creating an environment and installing NEoST's dependencies
+===========================================================
+
+NEoST was developed in Python 3. We recommend installing NEoST and its dependencies in a virtual environment. There are two main options for creating an environment and installing the dependencies: Python's first-party package installer, pip, and the third-party Conda package manager. Using conda may be easier because while conda can install MultiNest, pip cannot. The downside is that you will have to install conda.
+
+You need pip in both cases. It is likely already installed. If not, follow the official instructions for installing it.
+
+Using pip
+---------
+
+Two of NEoST's dependencies, GSL (GNU Scientific Library) and MultiNest, are not written in Python and are as such not available in the Python package index (PyPI). They need to be installed separately.
+
+GSL is widely available and can almost always be installed with your operating system's package manager, or loaded with something like ``module load gsl`` on a computing cluster.  Alternatively you can compile it yourself using the official instructions, see `Downloading GSL <https://www.gnu.org/software/gsl/#downloading>`_.
+
+However you choose to obtain GSL, make sure it works by executing
+
+.. code-block:: bash
+
+	gsl-config --version
+
+which should print the installed GSL version.
+
+MultiNest, on the other hand, must likely be compiled manually. Follow these steps (which assume that you have ``git``, a ``C++`` compiler, ``cmake``, and ``make`` installed):
+
+.. code-block:: bash
+
+	git clone https://github.com/JohannesBuchner/MultiNest
+	cd MultiNest/build
+	cmake ..
+	make
+
+You then need to add ``<path/to>/MultiNest/lib`` to an environment variable called ``LD_LIBRARY_PATH`` using
+
+.. code-block:: bash
+
+	export LD_LIBRARY_PATH=<path/to>/MultiNest/lib:$LD_LIBRARY_PATH
+
+Of course, exchange ``<path/to>`` for the actual path where you cloned MultiNest. You can add the export command to your ``$HOME/.bashrc`` file if you don't want to execute it in every new terminal in which you're running NEoST.
+
+The next step is to create a virtual environment and install NEoST. In Python, the official way to create a virtual environment is by using the venv module:
+
+.. code-block:: bash
+
+	python -m venv path/to/environment
+
+where you should substitute path/to/environment for a location of your choice. Then activate the environment with
+
+.. code-block:: bash
+
+	source path/to/environment/bin/activate
+
+To install NEoST itself, see Installing NEoST below.
+
+Using Conda
+-----------
+
+.. _basic_env:
+
+Assuming a Conda base environment has been installed and activated (see Conda installation instructions, e.g., `Miniconda <https://docs.anaconda.com/miniconda/>`_), a new virtual environment can then be created by navigating to the NEoST base directory and executing
+
+.. code-block:: bash
+
+	conda env create -f environment.yml
+
+This will create a new Conda environment called neost and install all dependencies (including GSL and MultiNest).  Once the environment has been created, activate it with
+
+.. code-block:: bash
+
+	conda activate neost
+
+This changes the active virtual environment from the default base
+environment to the new neost environment and needs to be entered any time
+NEoST is used.
+
[…]
