+++ _sources/contributing.rst.txt
@@ -1,0 +1,50 @@
+.. _contributing:
+
+Contributing
+------------
+
+We outline below the workflow we use for developing NEoST.
+
+Contact
+~~~~~~~
+
+You are welcome to contact members of the NEoST directly if you have
+any questions about the software or its use (for the current active list see
+the :ref:`acknowledgements` page).
+To get in touch with us regarding bugs and issues the easiest way is via the
+github `Issues <https://github.com/xpsi-group/neost/issues/>`_ page.
+
+Contributing
+~~~~~~~~~~~~
+
+To contribute to this project you can do so whilst being recognised as either
+a community member or a team member. If you contribute to feature development
+publicly on GitHub, you may well be invited to be recognised as a team member in
+the :ref:`acknowledgements`.
+
+Past developers are recognised as team members unless they no longer wish to be.
+Aside from acknowledging your contribution, this is useful for feature
+maintenance and as a contact point for questions, if permission is granted.
+Past team members will be acknowledged as such.
+
+If you wish to base a major research project on the
+development and/or application of NEoST we invite you to discuss you (and your
+group where applicable) joining the NEoST team in order to collaborate. If you
+find this idea interesting then please contact Anna Watts (contact details
+are on the :ref:`acknowledgements` page).
+
+.. _workflow:
+
+GitHub repository
+~~~~~~~~~~~~~~~~~
+
+NEoST is an open-source project and we aspire that all NEoST versions required
+for reproducibility of publications have release tags on GitHub.
+
+End-users and community contributors can interact with the project freely on
+GitHub. If you want to contribute a feature, you are welcome to communicate with us
+either on GitHub via issues and pull-requests, or by contacting a team member directly.
+
+
+
+

[…]
