### 2024-10-09 599fc51
Update Star.py to include ADM components
Specifically the ADM solve_structure and the functionality to call for specific components of the neutron stars, such ADM mass, radius, baryonic mass, and radius.
