+++ neost/Star.py
@@ -4,8 +4,10 @@ import warnings
 try:
     from neost.tovsolvers.TOVr import solveTOVr
     from neost.tovsolvers.TOVh import solveTOVh
+    from neost.tovsolvers.TOVdm import solveTOVdm
 except ImportError:
     from neost.tovsolvers.TOVr_python import solveTOVr
+    
     warnings.warn('C TOV solvers either not installed or broken, using Python TOV solver instead. This is much slower.')
 from . import global_imports
 
@@ -20,7 +22,7 @@ class Star():
     Instances of this class represent a model neutron star.
     """
 
-    def __init__(self, epscent, enthalpy=False):
+    def __init__(self, epscent,epscent_dm = 0.0, enthalpy=False):
         """
         Initialize a model star.
 
@@ -28,14 +30,22 @@ class Star():
             epscent (float): The central energy density of the
             model star in cgs units.
 
+            epscent_dm (float): The central energy density of the
+            dark matter component in cgs units
+
         """
         self.epscent = epscent
+        self.epscent_dm = epscent_dm
         self.Mb = 0
         self.Rns = 0
         self.tidal = 0
         self.enthalpy = enthalpy
+        self.radius_dm_core = 0
+        self.radius_dm_halo = 0
+        self.Mdm_core = 0
+        self.Mdm_halo = 0
 
-    def solve_structure(self, eps, pres, atol=1e-6, rtol=1e-4, hmax=1000., step=0.46):
+    def solve_structure(self, eps, pres,eps_dm = None,pres_dm = None,dm_halo = False,two_fluid_tidal = False, atol=1e-6, rtol=1e-4, hmax=1000., step=0.46):
         """
         Solve the relativistic structure equations to build a model star.
 
@@ -56,48 +66,111 @@ class Star():
         """
         
         if self.enthalpy is True:
-            self.Mb, self.Rns, self.tidal = solveTOVh(self.epscent,eps, pres)
+            if self.epscent_dm == 0.0:
+                self.Mb, self.Rns, self.tidal = solveTOVh(self.epscent,eps, pres)
+            else:
+                raise Exception("two-fluid enthalpy solve currently not available")
+               #two-fluid enthalpy tov solver not available currently
+               #self.Mb, self.Rns, self.tidal, self.Mdm_core, self.radius_dm_core = solveTOVdmh(self.epscent, self.epscent_dm, eps, pres, eps_dm, pres_dm) 
             
         else:
+            if self.epscent_dm == 0.0:
                 self.Mb, self.Rns, self.tidal, self.Gtt = solveTOVr(self.epscent, eps, pres, atol, rtol, hmax, step)
+                
+            else:
+                self.Mb, self.Rns, self.Mdm_core, self.Mdm_halo, self.radius_dm_core, self.radius_dm_halo, self.tidal = solveTOVdm(self.epscent, self.epscent_dm, eps, pres, eps_dm, pres_dm, dm_halo,two_fluid_tidal, atol, rtol, hmax, step)
 
 
-    @property
-    def Mrot(self):
-        """ Get the gravitational mass. """
-        return (self.Mb) / Msun 
+@property
+def Mrot(self):
+    """ Get the gravitational mass. """
+    return (self.Mb + self.Mdm_core + self.Mdm_halo) / Msun 
 
+@property
+def Mbaryonic(self):
+    """ Get the gravitational mass of baryonic component. """
+    return (self.Mb)/ Msun
 
     @property
[…]
