### 2024-10-09 d7f6d3a
Updates to PosteriorAnalysis.py to include ADM
