+++ neost/PosteriorAnalysis.py
@@ -3,7 +3,7 @@ import matplotlib
 from matplotlib.lines import Line2D
 import matplotlib.patches as mpatches
 from matplotlib.colors import ListedColormap
-import matplotlib.pyplot as plt
+from matplotlib import pyplot as plt
 import seaborn as sns
 from scipy.interpolate import UnivariateSpline
 from scipy.stats import gaussian_kde
@@ -60,15 +60,42 @@ def calc_bands(x, y):
     return miny, maxy
 
 def get_quantiles(array, quantiles=[0.025, 0.5, 0.975]):
-        contours = np.nanquantile(array, quantiles) #changed to nanquantile to inorder to ignore the nans that may appear
+        contours = np.nanquantile(array, quantiles)
         low = contours[0]
         median = contours[1]
         high = contours[2]
         minus = low - median
         plus = high - median
-        return np.round(median,2),np.round(plus,2),np.round(minus,2)  #returns uncertainties on the array
+        return np.round(median,2),np.round(plus,2),np.round(minus,2)
 
-def compute_table_data(root_name, EOS, variable_params, static_params):
+
+def compute_table_data(root_name, EOS, variable_params, static_params,dm = False):
+    """
+    Function to compute the table data in Raaijmakers et al. 2021 & Rutherford et al. 2024.
+    In particular: M_TOV, R_TOV, eps_cent_TOV, rho_cent_TOV, P_cent_TOV, R_1.4, eps_cent_1.4, rho_cent_1.4, P_cent_1.4,
+    R_2.0, eps_cent_2.0, rho_cent_2.0, P_cent_2.0. With v2.0.0, the user has the choice to include ADM or not.
+
+
+    Parameters
+    ----------
+
+    root_name: str
+        Name of the inference run to refer back to. Used to get the Multinest outputs.
+
+    EOS: obj
+        equation of state object initialized in the inference script, i.e., the parameters that are sampled during inferencing.
+
+    variable_params: dict
+        Variable parameters in the inference script.
+
+    static_params: dict
+        Static parameters in the inference script, i.e., the parameters that are held static during inference sampling.
+
+    dm: bool
+        If True ADM is included when computing the table data.
+
+
+    """
     ewposterior = np.loadtxt(root_name + 'post_equal_weights.dat')
     print("Total number of samples is %d" %(len(ewposterior)))
     try:
@@ -90,6 +117,9 @@ def compute_table_data(root_name, EOS, variable_params, static_params):
     except OSError:
         Data_array = np.zeros((len(ewposterior),13)) #contains Mtov, Rtov, eps_cent TOV, rho_cent TOV, P_cent TOV,R 1.4, eps_cent 1.4, rho_cent 1.4,
                                                                 #P_cent 1.4, R 2.0, eps_cent 2.0, rho_cent 2.0, P_cent 2.0
+                                                                #NOTE: ALL VALUES ARE THEIR ADMIXED VERSIONS WHEN dm == True!!
+
+
 
         for i in range(0, len(ewposterior), 1):
             pr = ewposterior[i][0:len(variable_params)]
@@ -98,44 +128,118 @@ def compute_table_data(root_name, EOS, variable_params, static_params):
             EOS.update(par, max_edsc=True)
 
             edsrho = UnivariateSpline(EOS.energydensities, EOS.massdensities, k=1, s=0)
[…]
