### 2024-10-09 d3c6f69
Update speedofsound.py to include ADM EoS
