### 2024-10-09 47a220e
Update tabulated.py to include ADM
