### 2024-10-09 d00ac7c
Update speedofsound.py to fix ADM EoS to match polytropes
