### 2024-10-09 3700e68
Examples using ADM in NEoST
