### 2024-10-09 35d6928
Updates to overview, starthere, and todo
All of this concerning the addition of dark matter into NEoST
