+++ neost/eos/tabulated.py
@@ -198,6 +198,6 @@ class TabulatedEoS(BaseEoS):
         else:
             check = False
             
-    return check
+        return check
 
 
