### 2024-10-09 829d235
Update Star.py
indentation fix
