### 2024-10-09 e2f3944
Tidier imports and removed trailing whitespaces 	modified:   neost/PosteriorAnalysis.py
