+++ neost/PosteriorAnalysis.py
@@ -1,9 +1,8 @@
 import numpy as np
-import matplotlib
 from matplotlib.lines import Line2D
 import matplotlib.patches as mpatches
 from matplotlib.colors import ListedColormap
-from matplotlib import pyplot as plt
+import matplotlib.pyplot as plt
 import seaborn as sns
 from scipy.interpolate import UnivariateSpline
 from scipy.stats import gaussian_kde
@@ -38,7 +37,7 @@ def m1(mc, m2):
 def calc_bands(x, y):
     miny = np.zeros((len(y),3))
     maxy = np.zeros((len(y),3))
-    
+
     for i in range(len(y)):
         z = y[i][y[i]>0.0]
         if len(z)<200:
@@ -54,7 +53,7 @@ def calc_bands(x, y):
         index_95 = np.argsort(array)[::-1][index_95]
         miny[i] =  x[i], min(testz[index_68]), min(testz[index_95])
         maxy[i] =  x[i], max(testz[index_68]), max(testz[index_95])
-        
+
     miny = miny[~np.all(miny == 0, axis=1)]
     maxy = maxy[~np.all(maxy == 0, axis=1)]
     return miny, maxy
@@ -180,13 +179,13 @@ def compute_table_data(root_name, EOS, variable_params, static_params,dm = False
                 Rdm = np.zeros(len(eps))
                 Mdm = np.zeros(len(eps))
 
-                
+
                 for j, e in enumerate(eps):
                     epsdm_cent = EOS.find_epsdm_cent(EOS.adm_fraction,eps)
                     epsdm[i] = epsdm_cent
                     star = Star(e,epsdm_cent)
-                    star.solve_structure(EOS.energydensities, EOS.pressures,EOS.energydensities_dm, EOS.pressures_dm, EOS.dm_halo) #EOS.two_fluid_tidal not needed in this section since only MR    
-                                                                                                                                    #so default value is used (False). 
+                    star.solve_structure(EOS.energydensities, EOS.pressures,EOS.energydensities_dm, EOS.pressures_dm, EOS.dm_halo) #EOS.two_fluid_tidal not needed in this section since only MR
+                                                                                                                                    #so default value is used (False).
                     M[j] = star.Mrot
                     R[j] = star.Req
                     Rdm[j] = star.Rdm
@@ -196,7 +195,7 @@ def compute_table_data(root_name, EOS, variable_params, static_params,dm = False
                 index_max_M = np.argmax(M[indices])
                 Radius_max_M = R[index_max_M]
 
-               
+
                 MR = UnivariateSpline(M, R[indices], k=1, s=0, ext=1)
                 epsdm_Mdm = UnivariateSpline(Mdm[indicies], epsdm[indices], k=1, s=0, ext=1)
 
@@ -236,7 +235,7 @@ def compute_table_data(root_name, EOS, variable_params, static_params,dm = False
 
                     pres_2 = EOS.eos(epsb_2) + eos_dm(epsdm_2)
 
-                
+
                 row = [max(M), Radius_max_M, np.log10(EOS.max_edsc + epsdm_max), max_rhoc, np.log10(EOS.eos(EOS.max_edsc) + eos_dm(epsdm_max)),R_14, np.log10(eps_14), rho_14, np.log10(pres_14),R_2, np.log10(eps_2), rho_2, np.log10(pres_2)]
 
 
@@ -313,9 +312,9 @@ def compute_prior_auxiliary_data(root_name, EOS, variable_params, static_params,
[…]
