### 2024-10-09 70b23d9
Fixed broken indentations 	modified:   neost/Star.py
