+++ neost/Star.py
@@ -65,113 +65,110 @@ class Star():
 
 
         """
-        
+
         if self.enthalpy is True:
             if self.epscent_dm == 0.0:
                 self.Mb, self.Rns, self.tidal = solveTOVh(self.epscent,eps, pres)
             else:
                 raise Exception("two-fluid enthalpy solve currently not available")
                #two-fluid enthalpy tov solver not available currently
-               #self.Mb, self.Rns, self.tidal, self.Mdm_core, self.radius_dm_core = solveTOVdmh(self.epscent, self.epscent_dm, eps, pres, eps_dm, pres_dm) 
-            
+               #self.Mb, self.Rns, self.tidal, self.Mdm_core, self.radius_dm_core = solveTOVdmh(self.epscent, self.epscent_dm, eps, pres, eps_dm, pres_dm)
+
         else:
             if self.epscent_dm == 0.0:
                 self.Mb, self.Rns, self.tidal, self.Gtt = solveTOVr(self.epscent, eps, pres, atol, rtol, hmax, step)
-                
+
             else:
                 self.Mb, self.Rns, self.Mdm_core, self.Mdm_halo, self.radius_dm_core, self.radius_dm_halo, self.tidal = solveTOVdm(self.epscent, self.epscent_dm, eps, pres, eps_dm, pres_dm, dm_halo,two_fluid_tidal, atol, rtol, hmax, step)
 
 
-@property
-def Mrot(self):
-    """ Get the gravitational mass. """
-    return (self.Mb + self.Mdm_core + self.Mdm_halo) / Msun 
+    @property
+    def Mrot(self):
+        """ Get the gravitational mass. """
+        return (self.Mb + self.Mdm_core + self.Mdm_halo) / Msun
 
-@property
-def Mbaryonic(self):
-    """ Get the gravitational mass of baryonic component. """
-    return (self.Mb)/ Msun
+    @property
+    def Mbaryonic(self):
+        """ Get the gravitational mass of baryonic component. """
+        return (self.Mb)/ Msun
 
     @property
     def Req(self):
         """ Get the equatorial radius. """
         return (self.Rns) / 1e5
-@property
-def Mdm(self):
-    """ Get the total gravitational dark matter mass. """
-    return (self.Mdm_core + self.Mdm_halo) / Msun
 
-@property
-def Mdmcore(self):
-    """ Get the dark matter mass within the baryonic radius. """
-    return (self.Mdm_core) / Msun
+    @property
+    def Mdm(self):
+        """ Get the total gravitational dark matter mass. """
+        return (self.Mdm_core + self.Mdm_halo) / Msun
 
-@property
-def Mdmhalo(self):
-    """ Get the dark matter mass outside the baryonic radius. """
-    return (self.Mdm_halo) / Msun
+    @property
+    def Mdmcore(self):
+        """ Get the dark matter mass within the baryonic radius. """
+        return (self.Mdm_core) / Msun
 
-@property
-def Rdm(self):
-    """ Get the total dark matter radius."""
-    if self.radius_dm_halo == 0 or self.radius_dm_halo == 999e5:
-        Rdm = self.radius_dm_core
-        
-    else:
-        Rdm = self.radius_dm_halo
-    return (Rdm) / 1e5
+    @property
+    def Mdmhalo(self):
+        """ Get the dark matter mass outside the baryonic radius. """
+        return (self.Mdm_halo) / Msun
 
-@property
-def Rdm_core(self):
[…]
