+++ running/plot_routines/MR_priors.py
@@ -1,0 +1,150 @@
+# Standard libraries
+import argparse
+import os
+
+# 3rd party
+import matplotlib.pyplot as plt
+import numpy as np
+
+# Local
+import plotting
+
+# Paths
+current_path = os.path.dirname(__file__)
+parent = os.path.dirname(current_path)
+
+# Input arguments
+parser = argparse.ArgumentParser()
+parser.add_argument('-n', '--numpy-histogram', action='store_true')
+parser.add_argument('-s', '--scatter', action='store_true')
+parser.add_argument('-t', '--tmp_output', action='store_true')
+parser.add_argument('-r', '--repro', action='store_true')
+args = parser.parse_args()
+
+# Choose plotting function (KDE, numpy-histogram, or scatterplot)
+plot = plotting.plot_KDE
+if args.numpy_histogram:
+    plot = plotting.plot_joint_distribution
+elif args.scatter:
+    plot = plotting.plot_scatter
+
+# Create subplots, remove spaces
+width = 7
+height = 5
+fig, axes = plt.subplots(nrows=2, ncols=2, sharex=True, sharey=False, figsize=(width,height))
+fig.subplots_adjust(wspace=0, hspace=0)
+
+# Set labels and titles
+xlabel = '$R$ [km]'
+ylabel = r'$M$ [$M_{\odot}$]'
+axes[1,0].set_xlabel(xlabel)
+axes[1,1].set_xlabel(xlabel)
+axes[0,0].set_ylabel(ylabel)
+axes[1,0].set_ylabel(ylabel)
+axes[0,0].set_title('PP')
+axes[0,1].set_title('CS')
+
+# Define colors
+c_n3lo = plotting.c_n3lo
+c_n2lo = plotting.c_n2lo
+c_hebeler = plotting.c_hebeler
+c_n3lo_11 = plotting.c_n3lo_11
+
+# Results directory
+directory = f'{parent}/results/prior' if not args.repro else f'{parent}/repro/prior'
+
+# PP
+# Keller N2LO, 1.1*rho_ns
+pp_n2lo_11 = np.loadtxt(f'{directory}/pp/n2lo/11/MR_prpr.txt')
+# Keller N2LO, 1.5*rho_ns
+pp_n2lo_15 = np.loadtxt(f'{directory}/pp/n2lo/15/MR_prpr.txt')
+# Keller N3LO, 1.1*rho_ns
+pp_n3lo_11 = np.loadtxt(f'{directory}/pp/n3lo/11/MR_prpr.txt')
+# Keller N3LO, 1.5*rho_ns
+pp_n3lo_15 = np.loadtxt(f'{directory}/pp/n3lo/15/MR_prpr.txt')
+# Hebeler
+pp_hebeler = np.loadtxt(f'{directory}/pp/hebeler/11/MR_prpr.txt')
+
+# CS
+# Keller N2LO, 1.1*rho_ns
+cs_n2lo_11 = np.loadtxt(f'{directory}/cs/n2lo/11/MR_prpr.txt')
+# Keller N2LO, 1.5*rho_ns
+cs_n2lo_15 = np.loadtxt(f'{directory}/cs/n2lo/15/MR_prpr.txt')
+# Keller N3LO, 1.1*rho_ns
+cs_n3lo_11 = np.loadtxt(f'{directory}/cs/n3lo/11/MR_prpr.txt')
+# Keller N3LO, 1.5*rho_ns
+cs_n3lo_15 = np.loadtxt(f'{directory}/cs/n3lo/15/MR_prpr.txt')
+# Hebeler
[…]
