### 2024-10-10 0ed4e7d
Modifications to the neost routines
	modified:   Likelihood.py
	modified:   PosteriorAnalysis.py
	modified:   Prior.py
	modified:   Star.py
	modified:   eos/base.py
	modified:   eos/polytropes.py
	modified:   eos/speedofsound.py
	modified:   utils.py
