+++ joss/paper.md
@@ -1,91 +1,0 @@
----
-title: 'NEoST: A Python package for nested sampling of the neutron star equation of state'
-tags:
-    - Python
-    - astrostatistics
-    - neutron stars
-authors:
-    - name: Geert Raaijmakers
-      orcid: 0000-0002-9397-786X
-      affiliation: 1
-    - name: Patrick Timmerman
-      orcid: 0009-0003-2793-1569
-      affiliation: 2
-    - name: Nathan Rutherford
-      orcid: 0000-0002-9626-7257
-      affiliation: 3
-    - name: Tuomo Salmi
-      orcid: 0000-0001-6356-125X
-      affiliation: 2
-    - name: Anna L. Watts
-      orcid: 0000-0002-1009-2354
-      affiliation: 2
-    - name:  Chanda Prescod-Weinstein
-      orcid: 0000-0002-6742-4532
-      affiliation: 3
-    - name:  Isak Svensson
-      orcid: 0000-0002-9211-5555
-      affiliation: 4
-    - name:  Melissa Mendes
-      orcid: 0000-0002-5250-0723
-      affiliation: 4
-affiliations:
-   - name: GRAPPA, Anton Pannekoek Institute for Astronomy and Institute of High-Energy Physics, University of Amsterdam, Science Park 904, 1098 XH Amsterdam, Netherlands
-     index: 1
-   - name: Anton Pannekoek Institute for Astronomy, University of Amsterdam, Science Park 904, 1098 XH Amsterdam, Netherlands
-     index: 2
-   - name: Department of Physics and Astronomy, University of New Hampshire, Durham, New Hampshire 03824, USA
-     index: 3
-   - name: Technische Universit\"at Darmstadt, Department of Physics, 64289 Darmstadt, Germany; ExtreMe Matter Institute EMMI, GSI Helmholtzzentrum f\"ur Schwerionenforschung GmbH, 64291 Darmstadt, Germany; Max-Planck-Institut f\"ur Kernphysik, Saupfercheckweg 1, 69117 Heidelberg, Germany
-     index: 4
-
-date: 5 September 2024
-bibliography: neostjoss.bib
----
-
-
-# Summary
-
-The Nested Equation of State Sampling (NEoST) package is an open-source code that allows users to infer the parameters of the dense matter Equation of State (EoS) in neutron stars via nested sampling. It provides a Bayesian inference framework that compares pre-existing EoS models (parameterized or tabulated, for both crust and core) to a variety of user-defined astrophysical input data (real or synthetic), namely mass-radius samples, mass-tidal deformability samples, and mass samples. NEoST can also be used to provide a fast solver for the Tolman-Oppenheimer-Volkoff equations for neutron star structure [@TOV1;@TOV2].  
-
-# Statement of need
-
-Matter in the cores of neutron stars can reach several times the nuclear saturation density. The EoS of matter under such circumstances is not well understood: in addition to extreme levels of neutron-richness there could also exist stable states of strange matter, in the form of either hyperons or deconfined quarks [@Hebeler:2015;@Lattimer:2016;@Tolos:2020]. Neutron star properties like mass, radius and tidal deformability depend on the EoS, so measurement of these quantities provides insight into the properties of ultradense nuclear matter.   
-
[…]
