+++ neost/eos/base.py
@@ -1,8 +1,8 @@
-import numpy as np
+import numpy
 from scipy.interpolate import UnivariateSpline
 from scipy import optimize
 from scipy.integrate import odeint
-import matplotlib.pyplot as plt
+from matplotlib import pyplot
 
 from .. Star import Star
 from .. import global_imports
@@ -18,8 +18,9 @@ gcm3_to_MeVfm3 = global_imports._gcm3_to_MeVfm3
 oneoverfm_MeV = global_imports._oneoverfm_MeV
 
 
-class BaseEoS():
-
+class BaseEoS():                       #### CAUTION: important alterations hard-coded to make NLO bands run
+                                       #### CAUTION: alterations in likelihood so that gaussian distribution in the ceft band, n3lo hard-coded
+                                       #### further alterations to sample gaussian in indexes as well
     """
     Base class representing an equation of state object.
 
@@ -50,10 +51,10 @@ class BaseEoS():
     def __init__(self, crust='ceft-Hebeler', rho_t=2e14):
 
         if crust not in ['ceft-Hebeler', 'ceft-Drischler', 'ceft-Lynn',
-                         'ceft-Tews', 'ceft-Keller-N2LO', 'ceft-Keller-N3LO', 'ceft-old', 'BPS', None]:
+                         'ceft-Tews', 'ceft-Keller-LO', 'ceft-Keller-NLO', 'ceft-Keller-N2LO', 'ceft-Keller-N3LO', 'ceft-Keller-N3LO-MOD', 'ceft-old', 'BPS', None]:
             raise TypeError('crust model not recognized, choose either \
                 "ceft-Hebeler", "ceft-Drischler", "ceft-Lynn", "ceft-Tews", \
-                "ceft-Keller-N2LO", "ceft-Keller-N3LO", "BPS" or None if no crust is needed')
+                "ceft-Keller-LO", "ceft-Keller-NLO", "ceft-Keller-N2LO", "ceft-Keller-N3LO", "ceft-Keller-N3LO-MOD", "BPS" or None if no crust is needed')
 
         self.crust = crust
         self.rho_t = rho_t
@@ -95,6 +96,30 @@ class BaseEoS():
                     self._rho_start_ceft = 0.5792
                     self._rho_end_BPS = 0.5
 
+                if crust == 'ceft-Keller-LO':
+                    self.min_norm = 1.624
+                    self.max_norm = 1.624000000001  #to mimic the existence of a max band
+                    self.min_index = 1.517
+                    self.max_index = 1.517
+                    self._rho_start_ceft = 0.5792
+                    self._rho_end_BPS = 0.5
+
+                #if crust == 'ceft-Keller-NLO-0.5':
+                #    self.min_norm = 0.548
+                #    self.max_norm = 1.286
+                #    self.min_index = 2.110
+                #    self.max_index = 1.729
+                #    self._rho_start_ceft = 0.5792
+                #    self._rho_end_BPS = 0.225
+
+                if crust == 'ceft-Keller-NLO':
+                    self.min_norm = 0.531
+                    self.max_norm = 1.283
+                    self.min_index = 2.229
+                    self.max_index = 1.738
+                    self._rho_start_ceft = 0.7
+                    self._rho_end_BPS = 0.35
+
                 if crust == 'ceft-Keller-N2LO':
                     self.min_norm = 1.81356
                     self.max_norm = 3.49759
@@ -111,6 +136,13 @@ class BaseEoS():
                     self._rho_start_ceft = 0.5792
                     self._rho_end_BPS = 0.5
 
[…]
