### 2024-10-16 #6003 @drannawatts
We have a fairly major update to NEoST (v2.0) coming that will add dark matter inference functionality, associated with this recent paper by @nr1118  (https://arxiv.org/abs/2410.00140).  Does it completely mess up your review process for v1.0 @adonath if we roll this out?  We don't want to make things unnecessarily difficult for you, but we're also keen to get it out there if we can.
