### 2024-10-22 134f7bd
Yet another ragged arrays issue, this one mainly appears in the unit tests and when UltraNest is used 	modified:   neost/eos/base.py
