### 2024-10-24 05d657d
Added instructions for creating an editable install. 	modified:   docs/source/install.rst
