### 2024-10-24 ebc6a0c
PosteriorAnalysis: Modified compute_auxiliary_data() and compute_prior_auxiliary_data() to accept both Multinest and Ultranest samples. Also reduced code duplication by introducing functions for loading the data and saving the output. 	modified:   neost/PosteriorAnalysis.py
