### 2024-10-25 b24fdde
Update applications.rst
Changed the APJ Letters articles to use the actual article link as opposed to the ADS link since the articles are open access. Other articles are kept as their ADS links since they are not open access.
