### 2024-11-05 #6003 @adonath
@drannawatts I think can live with the fact that part of the review was done on an older version of the package. However it is up to you which version you would like to describe in the paper and which development model you have. We should probably aim for concistency.  From the JOSS perspective I think we can handle both. There is no requirement that the version must be the most recent one. So the options are:

- Keep the paper mostly as is and describe v1.0. I will finish the review and comments must be back ported to the v1.0 branch. This would also require versioned docs, such that I can take a look at the respective documentation. Once the review is done you release v1.1 or v1.0.1
- Update the paper and describe v2.0. I would finish the review on the basis of the latest changes than. And ask @chaitanyaafle for any objections.

It seems to me option 2 is simpler.
