### 2024-11-06 a07f6dc
Update setup.py
including Isak's fix suggestion
