### 2024-11-06 #6003 @drannawatts
Agreed, Option 2 seems simpler!  We'll update the paper, we only need to modify a couple of lines.
