### 2024-11-06 ef2a330
Update ci_tests.yml
first step to including a macos M1 runner
