+++ .github/workflows/ci_tests.yml
@@ -49,4 +49,45 @@ jobs:
 
     - name: Test with pytest
       run: |
-        pytest        
+        pytest
+
+  tests-mac:
+    runs-on: macos-latest
+    defaults:
+      run:
+        shell: bash -l {0}
+    strategy:
+      fail-fast: false
+      max-parallel: 5
+      matrix:
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Install Conda environment with Micromamba
+      uses: mamba-org/setup-micromamba@v1
+      with:
+        environment-file: environment.yml
+        cache-env: true
+        cache-env-key: env-envkey-${{ github.sha }}-${{ github.run_attempt }}
+        environment-name: neost
+    - name: mamba installs
+      run: |
+        micromamba install flake8 pytest
+    - name: install gsl
+      run: |
+        sudo apt-get install -y libgsl-dev
+    - name: Lint with flake8
+      run: |
+        # stop the build if there are Python syntax errors or undefined names
+        #flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: install neost package
+      run: |
+        pip install .
+      shell: bash -l {0}
+
+    - name: Test with pytest
+      run: |
+        pytest
