### 2024-11-06 11e5480
Update ci_tests.yml
changing gsl install to the mac way of doing it with brew
