### 2024-11-06 478afa0
Update ci_tests.yml
Seems there is a need to remove the -l argument otherwise the macos runner will always use Python 2.7.18. See this thread for further details: https://github.com/actions/setup-python/issues/179#issuecomment-750495666
