### 2024-11-06 e4a7030
Update ci_tests.yml
Just checking the Python version
