### 2024-11-06 e2850dd
Update setup.py
Forgot to add include_dirs = tuple(include_dirs) for the Mac stuff
