### 2024-11-07 4c79987
Update ci_tests.yml
Looking at the issue thread with Axel, it appears that we also might need to install pymultinest with pip, thus also needing to create a new enviroment.yml file, which has pymultinest deleted?
