+++ joss/neostjoss.bib
@@ -52,6 +52,25 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Nelson:2018,
+       author = {{Nelson}, Ann E. and {Reddy}, Sanjay and {Zhou}, Dake},
+        title = "{Dark halos around neutron stars and gravitational waves}",
+      journal = {\jcap},
+     keywords = {High Energy Physics - Phenomenology},
+         year = 2019,
+        month = jul,
+       volume = {2019},
+       number = {7},
+          eid = {012},
+        pages = {012},
+          doi = {10.1088/1475-7516/2019/07/012},
+archivePrefix = {arXiv},
+       eprint = {1803.03266},
+ primaryClass = {hep-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019JCAP...07..012N},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 
 
 @ARTICLE{Raaijmakers:2019,
@@ -149,6 +168,23 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Rutherford:2024b,
+       author = {{Rutherford}, Nathan and {Prescod-Weinstein}, Chanda and {Watts}, Anna},
+        title = "{Probing fermionic asymmetric dark matter cores using global neutron star properties}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics, High Energy Physics - Phenomenology, Nuclear Theory},
+         year = 2024,
+        month = sep,
+          eid = {arXiv:2410.00140},
+        pages = {arXiv:2410.00140},
+          doi = {10.48550/arXiv.2410.00140},
+archivePrefix = {arXiv},
+       eprint = {2410.00140},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv241000140R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Choudhury:2024,
        author = {{Choudhury}, Devarshi and {Salmi}, Tuomo and {Vinciguerra}, Serena and {Riley}, Thomas E. and {Kini}, Yves and {Watts}, Anna L. and {Dorsman}, Bas and {Bogdanov}, Slavko and {Guillot}, Sebastien and {Ray}, Paul S. and {Reardon}, Daniel J. and {Remillard}, Ronald A. and {Bilous}, Anna V. and {Huppenkothen}, Daniela and {Lattimer}, James M. and {Rutherford}, Nathan and {Arzoumanian}, Zaven and {Gendreau}, Keith C. and {Morsink}, Sharon M. and {Ho}, Wynn C.~G.},
         title = "{A NICER View of the Nearest and Brightest Millisecond Pulsar: PSR J0437-4715}",
@@ -187,6 +223,62 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2022ApJ...941..150S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Salmi:2023,
[…]
