### 2024-11-08 ce8b9dc
Update neostjoss.bib
journal shortcut updates changed to full journal names
