+++ joss/neostjoss.bib
@@ -55,7 +55,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Nelson:2018,
        author = {{Nelson}, Ann E. and {Reddy}, Sanjay and {Zhou}, Dake},
         title = "{Dark halos around neutron stars and gravitational waves}",
-      journal = {\jcap},
+      journal = {Journal of Cosmology and Astroparticle Physics},
      keywords = {High Energy Physics - Phenomenology},
          year = 2019,
         month = jul,
@@ -227,7 +227,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Salmi:2023,
        author = {{Salmi}, Tuomo and {Vinciguerra}, Serena and {Choudhury}, Devarshi and {Watts}, Anna L. and {Ho}, Wynn C.~G. and {Guillot}, Sebastien and {Kini}, Yves and {Dorsman}, Bas and {Morsink}, Sharon M. and {Bogdanov}, Slavko},
         title = "{Atmospheric Effects on Neutron Star Parameter Constraints with NICER}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {Neutron stars, X-ray astronomy, 1108, 1810, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics, Nuclear Theory},
          year = 2023,
         month = oct,
@@ -246,7 +246,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Salmi:2024a,
        author = {{Salmi}, Tuomo and {Choudhury}, Devarshi and {Kini}, Yves and {Riley}, Thomas E. and {Vinciguerra}, Serena and {Watts}, Anna L. and {Wolff}, Michael T. and {Arzoumanian}, Zaven and {Bogdanov}, Slavko and {Chakrabarty}, Deepto and {Gendreau}, Keith and {Guillot}, Sebastien and {Ho}, Wynn C.~G. and {Huppenkothen}, Daniela and {Ludlam}, Renee M. and {Morsink}, Sharon M. and {Ray}, Paul S.},
         title = "{The Radius of the High-mass Pulsar PSR J0740+6620 with 3.6 yr of NICER Data}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {Neutron stars, X-ray astronomy, 1108, 1810, Astrophysics - High Energy Astrophysical Phenomena, Nuclear Theory},
          year = 2024,
         month = oct,
