### 2024-11-13 37fee2a
Update paper.md
changed a comma to a semi-colon
