### 2024-11-14 ac07d07
Update install.rst to use python -m pip install
