### 2024-11-15 fc2fea5
Update ci_tests.yml
changed to brew install flake8 pytest
