+++ .github/workflows/ci_tests.yml
@@ -55,7 +55,7 @@ jobs:
     runs-on: macos-latest
     defaults:
       run:
-        shell: bash {0}
+        shell: bash -l {0}
     strategy:
       fail-fast: false
       max-parallel: 5
@@ -84,7 +84,7 @@ jobs:
         echo $PATH
         python --version
         which python
-      shell: bash {0}
+      shell: bash -l {0}
     
     - name: install gsl
       run: |
@@ -98,7 +98,7 @@ jobs:
     - name: install neost package
       run: |
         pip install .
-        shell: bash {0}
+        shell: bash -l {0}
 
     - name: Test with pytest
       run: |
