+++ .github/workflows/ci_tests.yml
@@ -102,4 +102,4 @@ jobs:
 
     - name: Test with pytest
       run: |
-        pytest 
+        python -m pytest 
