### 2024-11-28 06b0c01
Change compute_prior_auxiliary_data() to be MPI parallelized.
Main downside at the moment is that sorting is changed compared to the equally-weighted samples.
	modified:   PosteriorAnalysis.py
