+++ neost/PosteriorAnalysis.py
@@ -17,6 +17,9 @@ from neost.Star import Star
 from neost.Likelihood import Likelihood
 import neost.global_imports as global_imports
 
+# MPI
+from mpi4py import MPI
+
 c = global_imports._c
 G = global_imports._G
 Msun = global_imports._M_s
@@ -191,36 +194,47 @@ def save_auxiliary_data(path, identifier, data, fnames):
         print(f'Writing {fname} to disk')
 
 def compute_prior_auxiliary_data(path, EOS, variable_params, static_params, sampler='ultranest', identifier=''):
-    ewprior = load_equal_weighted_samples(path, sampler, identifier)
-    print(f"Total number of samples is {len(ewprior)}")
-
-    num_stars = len(np.array([v for k,v in variable_params.items() if 'rhoc' in k]))
-
-    if len(list(variable_params.keys())) == num_stars:
-        flag = True
-
-    else:
-        flag = False
+    comm = MPI.COMM_WORLD
+    mpi_rank = comm.Get_rank()
+    num_processes = comm.Get_size()
+    samples = [[] for i in range(num_processes)] # This is essentially a rearranged 'ewprior'
 
     masses = np.linspace(.2, 2.9, 50)
     energydensities = np.logspace(14.2, 16, 50)
+    pressures = None
+    pressures_rho = None
+    flag = None
 
-    if flag == True:
+    if mpi_rank == 0:
+        ewprior = load_equal_weighted_samples(path, sampler, identifier)
+        num_stars = len(np.array([v for k,v in variable_params.items() if 'rhoc' in k]))
+        print(f"Total number of samples is {len(ewprior)}, and the number of stars is {num_stars}")
+
+        if len(list(variable_params.keys())) == num_stars:
+            flag = True
+        else:
+            flag = False
+
+        # Final results will be gathered in these later (after the loop), for now just create them
         pressures = np.zeros((len(masses), len(ewprior)))
         pressures_rho = np.zeros((len(masses), len(ewprior)))
-        MR_prpr_pp = np.zeros((len(ewprior), 2))
-    else:
-        pressures = np.zeros((len(masses), len(ewprior)))
-        pressures_rho = np.zeros((len(masses), len(ewprior)))
-        minpres = np.zeros((3, len(energydensities)))
-        maxpres = np.zeros((3, len(energydensities)))
-        minpres_rho = np.zeros((3, len(energydensities)))
-        maxpres_rho = np.zeros((3, len(energydensities)))
-        MR_prpr_pp = np.zeros((len(ewprior), 2))
 
-    for i in range(0, len(ewprior), 1):
+        # Recast ewprior in a form suitable for MPI scatter
+        for i in range(len(ewprior)):
+            samples[i%num_processes].append(ewprior[i])
 
-        pr = ewprior[i][0:len(variable_params)]
+    # Scatter samples to the different processes
+    samples = comm.scatter(samples, root=0)
+
+    # These three are now local to each process, and each will be joined together with its siblings later (with comm.gather)
+    mass_radius = [] # Calling it mass_radius because MR_prpr_pp is a bad name
+    pressures_list = [] # These two are a little complicated, so using the suffix _list to clarify that this is not the final output
+    pressures_rho_list = []
+    pressures_indexes = []
+    pressures_rho_indexes = []
+
+    for sample in samples:
+        pr = sample[0:len(variable_params)]
         par = {e:pr[j] for j, e in enumerate(list(variable_params.keys()))}
         par.update(static_params)
         EOS.update(par, max_edsc=True)
[…]
