### 2024-11-28 8c8efb7
Simplified and improved how pressures.npy is handled in the MPI version of compute_prior_auxiliary_data. 	modified:   PosteriorAnalysis.py
