### 2024-11-28 02c728c
compute_prior_auxiliary_data(): change name of "flag" to "eos_is_fixed" 	modified:   PosteriorAnalysis.py
