### 2024-12-05 660aac6
Update setup.py
following Axel's suggestion
