+++ .github/workflows/ci_tests.yml
@@ -98,7 +98,6 @@ jobs:
     - name: install neost package
       run: |
         pip install .
-        shell: bash -l {0}
 
     - name: Test with pytest
       run: |
