### 2024-12-05 #6003 @nr1118
Quick question @adonath, in point #2 of your message last week, did you mean, "clarify that PyMultinest is not supported," ?
