+++ .github/workflows/ci_tests.yml
@@ -99,6 +99,8 @@ jobs:
       run: |
         pip install .
 
+      shell: bash -l {0}
+
     - name: Test with pytest
       run: |
-        python -m pytest 
+        python -m pytest
