### 2024-12-05 #6003 @adonath
@nr1118 Sorry, I there was a typo in my comment. Autocorrect changed Cython to Python. What I meant was clarifying that Cython is not supported on M1. From what I have tried, the compilation could not succeed without manual adjustment by the user.
