### 2024-12-05 #6003 @nr1118
@adonath Ah, I see. Although, even with the manual adjustment that you suggested, I am seeing that in the MAC runner that I made (see Fixes to issue #48 branch), I cannot get the TOV solvers to cythonize and it always uses the python solvers. I take it that is the issue you are running into?
