### 2024-12-05 501e5aa
Update setup.py
swapped appends for extends for OS users
