+++ setup.py
@@ -42,8 +42,8 @@ if 'darwin' in OS:
     # os.environ["CXX"] = "/usr/local/opt/llvm/bin/clang++"
     library_dirs.extend('/usr/local/opt/llvm/lib')
     library_dirs.extend('/opt/local/lib')
-    extra_compile_args.extend('-Wno-#warnings')
-    extra_compile_args.extend('-Wno-error=format-security')
+    extra_compile_args.append('-Wno-#warnings')
+    extra_compile_args.append('-Wno-error=format-security')
     include_dirs.extend(['/usr/local/include', '/usr/local/opt/llvm/include', './neost/tovsolvers/'])
 
 else:
