### 2024-12-05 67bfedd
Update setup.py
changing append to extend did nothing
