### 2024-12-05 a8a588e
Update test_likelihood.py
