### 2024-12-05 fe00b80
Update test_likelihood.py
