### 2024-12-05 7ef068c
Update test_likelihood.py
