+++ tests/test_likelihood.py
@@ -36,8 +36,8 @@ class TestLikelihoodCheck(object):
         ll = likelihood.call(par)
         print(ll)
 	if np.isclose(-4.187612429216115, ll, rtol=1.0e-5) is False:
-		print('C TOV solvers either not installed or broken, using Python TOV solver instead')
-		assert np.isclose(--4.192614304543445, ll, rtol=1.0e-5)
+	    print('C TOV solvers either not installed or broken, using Python TOV solver instead')
+	    assert np.isclose(--4.192614304543445, ll, rtol=1.0e-5)
 	else:
-		assert np.isclose(-4.187612429216115, ll, rtol=1.0e-5)
+	    assert np.isclose(-4.187612429216115, ll, rtol=1.0e-5)
 
