+++ tests/test_likelihood.py
@@ -35,9 +35,10 @@ class TestLikelihoodCheck(object):
         par = prior.inverse_sample(cube)
         ll = likelihood.call(par)
         print(ll)
-	if np.isclose(-4.187612429216115, ll, rtol=1.0e-5) is False:
-	    print('C TOV solvers either not installed or broken, using Python TOV solver instead')
-	    assert np.isclose(--4.192614304543445, ll, rtol=1.0e-5)
-	else:
-	    assert np.isclose(-4.187612429216115, ll, rtol=1.0e-5)
+        if np.isclose(-4.187612429216115, ll, rtol=1.0e-5) is False:
+            print('C TOV solvers either not installed or broken, using Python TOV solver instead')
+            assert np.isclose(--4.192614304543445, ll, rtol=1.0e-5)
+            
+        else:
+            assert np.isclose(-4.187612429216115, ll, rtol=1.0e-5)
 
