+++ _sources/Start_here.rst.txt
@@ -1,0 +1,27 @@
+.. _landing_page_tutorials:
+
+==========
+Start here
+==========
+
+All tutorials listed below may be found as Jupyter notebooks under ``docs/source/``. We encourage running these yourself as you go through the tutorials as a means of practice. Note, these scripts can take a considerable amount of time to run on your personal computer. To shorten the time to run the scripts, we recommend changing ``n_live_points=5000`` to ``n_live_points=500``. Any external data files needed such as the AP4 tabulated equation of state data and mass-radius data are available on `GitHub <https://github.com/xpsi-group/neost/tree/main/examples>`_. These data files should be saved in ``neost/examples/``.
+
+**Overview of the available tutorials:**
+
+* :doc:`Piecewise Polytropic Example<PP_example>` a python script detailing how to perform a basic inference run using the piecewise polytropic equation of state parameterization.
+
+* :doc:`Piecewise Polytropic Prior Example<PP_prior_example>` a python script detailing how to perform prior sampling using the piecewise polytropic equation of state parameterization.
+
+* :doc:`Piecewise Polytropic with ADM Example<PP_with_ADM_example>` a python script detailing how to perform a basic inference run using the piecewise polytropic equation of state parameterization and the Nelson et al. 2018 bosonic asymmetric dark matter core model.
+
+* :doc:`Piecewise Polytropic with ADM Prior Example<PP_with_ADM_prior_example>` a python script detailing how to perform prior sampling using the piecewise polytropic equation of state parameterization and the Nelson et al. 2018 bosonic asymmetric dark matter core model.
+
+* :doc:`ADM admixed M-R and M-Tidal Example<ADM_MR_Tidal_tutorial>` a python script detailing how to compute the mass-radius and mass-tidal deformability curve when ADM is considered.
+
+* :doc:`Speed of Sound Example<CS_example>` a python script demonstrating how to perform a basic inference run using the speed of sound equation of state parameterization.
+
+* :doc:`Speed of Sound Prior Example Example<CS_prior_example>` a python script detailing how to perform prior sampling using the speed of sound equation of state parameterization.
+
+* :doc:`Tabulated Example<Tabulated_example>` a python script demonstrating how to perform an inference run using the AP4 tabulated equation of state.
+
+* :doc:`Tabulated Prior Example<Tabulated_prior_example>` an informative python script detailing how to perform a basic prior sampling run using the AP4 tabulated equation of state.
