### 2024-12-06 bc6af45
Update ci_tests.yml
adding brew install cython to see if this changes anything
