### 2024-12-06 424b2be
Update environment_mac.yml
