### 2024-12-06 31b931c
Update setup.py to match main
