### 2024-12-06 da4c924
Update ci_tests.yml to install cython w/ pip
