### 2024-12-13 892040c
Cython versus Python statement
Included approximate computational speed-ups from using Cython over Python. This was done by using %timeit -r 10 magic function to  compute approximate run times.
