### 2024-12-16 #6003 @nr1118
@adonath I have updated the documentation to clarify that Cython is not supported for MAC M-series users (see PR #67) and I have also inserted a quantifiable number on the relative speed-up of going from Python TOV solvers to Cython TOV solvers as well (as it seems you have seen given that you liked my closing comment).
