### 2024-12-16 #6003 @adonath
@drannawatts and @nr1118 Are the paper and author list up to date after the changes introduced in version 2.0?
