### 2024-12-17 #6003 @adonath
@drannawatts I have no additional comments on the paper, so you can proceed to do a new release of the package and upload to an archive of your choice. Please make sure the archive title, author list, version tag and license are consistent. Finally please share the version and DOI here in the thread. Thanks!
