### 2024-12-18 21326eb
Final changes for JOSS release
