+++ CHANGELOG.rst
@@ -33,6 +33,28 @@ and this project adheres to
 .. Attribution
 .. ^^^^^^^^^^^
 
+[v2.1.0 - 2024-12-18]
+~~~~~~~~~~~~~~~~~~~~~
+Summary
+^^^^^^^
+Changes from JOSS review process. 
+
+Added
+^^^^^
+*Notes for those attempting to compile and run on Mac M-series chips
+*Details of Cython vs Python speed-up
+
+Changed
+^^^^^^^
+* Small changes to JOSS paper
+* Updated install.rst to use python -m pip install
+
+Attribution
+^^^^^^^^^^^
+* NEoST core team
+* Axel Donath
+
+
 [v2.0.0 - 2024-10-01]
 ~~~~~~~~~~~~~~~~~~~~~
 Summary
