### 2024-12-18 a575736
Final changes +1 for JOSS release
