### 2024-12-19 3580995
Update CHANGELOG.rst
