### 2024-12-19 29fdd5d
Update PosteriorAnalysis.py
Changed language to include future versions of NEoST and this feature
