+++ _sources/ADM_MR_Tidal_tutorial.ipynb.txt
@@ -1,0 +1,231 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "80c768da-a30a-4829-92a3-983041613824",
+   "metadata": {},
+   "source": [
+    "# ADM M-R and M-Tidal Example\n",
+    "\n",
+    "To run this tutorial, you should install NEoST following the install guide. \n",
+    "\n",
+    "The following block of code will properly import NEoST and its prerequisites, furthermore it also defines how to compute the mass-radius and mass-tidal deformability relations with ADM."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "id": "9026c71d-b2dc-4a91-8585-7a7e09841eb3",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import time\n",
+    "import neost\n",
+    "from neost.eos import polytropes\n",
+    "from neost.Prior import Prior\n",
+    "from neost.Star import Star\n",
+    "from neost.Likelihood import Likelihood\n",
+    "from scipy.stats import multivariate_normal\n",
+    "from neost import PosteriorAnalysis\n",
+    "import numpy\n",
+    "import matplotlib\n",
+    "from matplotlib import pyplot\n",
+    "import timeit\n",
+    "import neost.global_imports as global_imports\n",
+    "from matplotlib.patches import Ellipse\n",
+    "\n",
+    "start = time.time()\n",
+    "c = global_imports._c\n",
+    "G = global_imports._G\n",
+    "Msun = global_imports._M_s\n",
+    "pi = global_imports._pi\n",
+    "rho_ns = global_imports._rhons"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "id": "130737b6-b6fe-47bb-8dd9-ef5022447e0d",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "EOS = polytropes.PolytropicEoS(crust='ceft-Hebeler', rho_t= 2e14,adm_type = 'Fermionic',dm_halo = True,two_fluid_tidal = True)"
+   ]
+  },
+  {
[…]
