### 2024-12-19 #6003 @drannawatts
Version number for the JOSS release is v2.1.0

Zenodo doi is https://doi.org/10.5281/zenodo.14530057
