### 2024-12-19 #6003 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
