### 2025-01-06 b25159d
Initial editorial tweaks for JOSS publication
