+++ joss/paper.md
@@ -25,10 +25,10 @@ authors:
       affiliation: 3
     - name:  Isak Svensson
       orcid: 0000-0002-9211-5555
-      affiliation: 4
+      affiliation: "4, 5, 6"
     - name:  Melissa Mendes
       orcid: 0000-0002-5250-0723
-      affiliation: 4
+      affiliation: "4, 5, 6"
 affiliations:
    - name: GRAPPA, Anton Pannekoek Institute for Astronomy and Institute of High-Energy Physics, University of Amsterdam, Science Park 904, 1098 XH Amsterdam, Netherlands
      index: 1
@@ -36,8 +36,12 @@ affiliations:
      index: 2
    - name: Department of Physics and Astronomy, University of New Hampshire, Durham, New Hampshire 03824, USA
      index: 3
-   - name: Technische Universit\"at Darmstadt, Department of Physics, 64289 Darmstadt, Germany; ExtreMe Matter Institute EMMI, GSI Helmholtzzentrum f\"ur Schwerionenforschung GmbH, 64291 Darmstadt, Germany; Max-Planck-Institut f\"ur Kernphysik, Saupfercheckweg 1, 69117 Heidelberg, Germany
+   - name: Technische Universität Darmstadt, Department of Physics, 64289 Darmstadt, Germany
      index: 4
+   - name: ExtreMe Matter Institute EMMI, GSI Helmholtzzentrum für Schwerionenforschung GmbH, 64291 Darmstadt, Germany
+     index: 5
+   - name: Max-Planck-Institut für Kernphysik, Saupfercheckweg 1, 69117 Heidelberg, Germany
+     index: 6
 
 date: 5 September 2024
 bibliography: neostjoss.bib
@@ -46,17 +50,17 @@ bibliography: neostjoss.bib
 
 # Summary
 
-The Nested Equation of State Sampling (NEoST) package is an open-source code that allows users to infer the parameters of the dense matter Equation of State (EoS) in neutron stars via nested sampling. It provides a Bayesian inference framework that compares pre-existing EoS models (parameterized or tabulated, for both crust and core) to a variety of user-defined astrophysical input data (real or synthetic), namely mass-radius samples, mass-tidal deformability samples, and mass samples. NEoST can also be used to provide a fast solver for the Tolman-Oppenheimer-Volkoff (TOV) equations for neutron star structure [@TOV1;@TOV2]. Moreoever, NEoST is able to fully account for a possible dark matter component inside neutron stars, where the dark matter is described by the [@Nelson:2018] model.   
+The Nested Equation of State Sampling (NEoST) package is an open-source code that allows users to infer the parameters of the dense matter Equation of State (EoS) in neutron stars via nested sampling. It provides a Bayesian inference framework that compares pre-existing EoS models (parameterized or tabulated, for both crust and core) to a variety of user-defined astrophysical input data (real or synthetic), namely mass-radius samples, mass-tidal deformability samples, and mass samples. NEoST can also be used to provide a fast solver for the Tolman--Oppenheimer--Volkoff (TOV) equations for neutron star structure [@TOV1;@TOV2]. Moreover, NEoST is able to fully account for a possible dark matter component inside neutron stars, where the dark matter is described by the @Nelson:2018 model.
 
 # Statement of need
 
-Matter in the cores of neutron stars can reach several times the nuclear saturation density. The EoS of matter under such circumstances is not well understood: in addition to extreme levels of neutron-richness there could also exist stable states of strange matter, in the form of either hyperons or deconfined quarks [@Hebeler:2015;@Lattimer:2016;@Tolos:2020]. Neutron star properties like mass, radius and tidal deformability depend on the EoS, so measurement of these quantities provides insight into the properties of ultradense nuclear matter.   
[…]
