### 2025-01-06 #6003 @warrickball
Hi @drannawatts — I've just opened [a small PR](https://github.com/xpsi-group/neost/pull/69) with some editorial tweaks. Note the comment about affiliation number 1, which should be split up if it's actually more than one affiliation.

Once that's merged I'll give the publication pipeline another test run and check the metadata.
