### 2025-01-07 36ed590
Update paper.md
Thanks for the changes! 

Affiliation 1 should be shown as a single affiliation (GRAPPA is a composite institute with 3 possible sub-components, so the way that - apparently - they denote things is to give first GRAPPA and then the sub-strands). 

I also realised in doing this that we needed to fix the ordering of the institutes since we had flipped author ordering at some point, so I've swapped 2 and 3.
