+++ joss/neostjoss.bib
@@ -529,7 +529,7 @@ archivePrefix = {arXiv},
 @ARTICLE{MultiNest_2009,
     author = {{Feroz}, F. and {Hobson}, M.~P. and {Bridges}, M.},
     title = "{MULTINEST: an efficient and robust Bayesian inference tool for cosmology and particle physics}",
-    journal = {Monthly Notices of the RAS},
+    journal = {Monthly Notices of the Royal Astronomical Society},
     archivePrefix = "arXiv",
     eprint = {0809.3437},
     keywords = {methods: data analysis , methods: statistical},
