### 2025-01-07 #6003 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
