### 2025-01-07 #6003 @warrickball
Thanks once again to @chaitanyaafle for reviewing and @adonath for editing and reviewing this submission! JOSS simply wouldn't be possible without its community of volunteers.
