### 2025-01-08 #6003 @drannawatts
Thank you so much @adonath @chaitanyaafle @warrickball (and the JOSS Editorial Bot of course!).
